Refactor GA logic into separate mixin
