+++ codes/mainwindow.py
@@ -50,6 +50,7 @@ from gui.menu_mixin import MenuMixin
 from gui.beam_mixin import ContinuousBeamMixin
 from gui.microchip_mixin import MicrochipPageMixin
 from gui.main_window.theme_mixin import ThemeMixin
+from gui.main_window.ga_mixin import GAOptimizationMixin
 
 
 # Additional libraries used
@@ -80,7 +81,7 @@ except ImportError:
 
         
 
-class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin, ThemeMixin):
+class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin, ThemeMixin, GAOptimizationMixin):
     def __init__(self):
         super().__init__()
         self.setWindowTitle("DeVana")
@@ -1594,412 +1595,6 @@ class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin
         layout.addWidget(self.sobol_sub_tabs)
         self.sobol_tab.setLayout(layout)
 
-    def create_ga_tab(self):
-        """Create the genetic algorithm optimization tab"""
-        self.ga_tab = QWidget()
-        layout = QVBoxLayout(self.ga_tab)
-
-        # Create sub-tabs widget
-        self.ga_sub_tabs = QTabWidget()
-
-        # -------------------- Sub-tab 1: GA Hyperparameters --------------------
-        ga_hyper_tab = QWidget()
-        ga_hyper_layout = QFormLayout(ga_hyper_tab)
-
-        self.ga_pop_size_box = QSpinBox()
-        self.ga_pop_size_box.setRange(1, 10000)
-        self.ga_pop_size_box.setValue(800)
-
-        self.ga_num_generations_box = QSpinBox()
-        self.ga_num_generations_box.setRange(1, 10000)
-        self.ga_num_generations_box.setValue(100)
-
-        self.ga_cxpb_box = QDoubleSpinBox()
-        self.ga_cxpb_box.setRange(0, 1)
-        self.ga_cxpb_box.setValue(0.7)
-        self.ga_cxpb_box.setDecimals(3)
-
-        self.ga_mutpb_box = QDoubleSpinBox()
-        self.ga_mutpb_box.setRange(0, 1)
-        self.ga_mutpb_box.setValue(0.2)
-        self.ga_mutpb_box.setDecimals(3)
-
-        self.ga_tol_box = QDoubleSpinBox()
-        self.ga_tol_box.setRange(0, 1e6)
-        self.ga_tol_box.setValue(1e-3)
-        self.ga_tol_box.setDecimals(6)
-
-        self.ga_alpha_box = QDoubleSpinBox()
-        self.ga_alpha_box.setRange(0.0, 10.0)
-        self.ga_alpha_box.setDecimals(4)
-        self.ga_alpha_box.setSingleStep(0.01)
-        self.ga_alpha_box.setValue(0.01)
-        
-        # Add benchmarking runs box
-        self.ga_benchmark_runs_box = QSpinBox()
-        self.ga_benchmark_runs_box.setRange(1, 1000)
-        self.ga_benchmark_runs_box.setValue(1)
-        self.ga_benchmark_runs_box.setToolTip("Number of times to run the GA for benchmarking (1 = single run)")
-        
-        # Add adaptive rates checkbox
-        self.adaptive_rates_checkbox = QCheckBox("Use Adaptive Rates")
-        self.adaptive_rates_checkbox.setChecked(False)
-        self.adaptive_rates_checkbox.setToolTip("Automatically adjust crossover and mutation rates during optimization")
-        self.adaptive_rates_checkbox.stateChanged.connect(self.toggle_adaptive_rates_options)
-        
-        # Create a widget to hold adaptive rate options
-        self.adaptive_rates_options = QWidget()
-        adaptive_options_layout = QFormLayout(self.adaptive_rates_options)
-        adaptive_options_layout.setContentsMargins(20, 0, 0, 0)  # Add left margin for indentation
-        
-        # Stagnation limit spinner
-        self.stagnation_limit_box = QSpinBox()
-        self.stagnation_limit_box.setRange(1, 50)
-        self.stagnation_limit_box.setValue(5)
-        self.stagnation_limit_box.setToolTip("Number of generations without improvement before adapting rates")
-        adaptive_options_layout.addRow("Stagnation Limit:", self.stagnation_limit_box)
-        
-        # Create a widget for crossover bounds
-        crossover_bounds_widget = QWidget()
-        crossover_bounds_layout = QHBoxLayout(crossover_bounds_widget)
-        crossover_bounds_layout.setContentsMargins(0, 0, 0, 0)
-        
-        self.cxpb_min_box = QDoubleSpinBox()
-        self.cxpb_min_box.setRange(0.01, 0.5)
-        self.cxpb_min_box.setValue(0.1)
-        self.cxpb_min_box.setDecimals(2)
-        self.cxpb_min_box.setSingleStep(0.05)
-        self.cxpb_min_box.setToolTip("Minimum crossover probability")
-        
-        self.cxpb_max_box = QDoubleSpinBox()
-        self.cxpb_max_box.setRange(0.5, 1.0)
-        self.cxpb_max_box.setValue(0.9)
-        self.cxpb_max_box.setDecimals(2)
-        self.cxpb_max_box.setSingleStep(0.05)
-        self.cxpb_max_box.setToolTip("Maximum crossover probability")
-        
-        crossover_bounds_layout.addWidget(QLabel("Min:"))
-        crossover_bounds_layout.addWidget(self.cxpb_min_box)
-        crossover_bounds_layout.addWidget(QLabel("Max:"))
-        crossover_bounds_layout.addWidget(self.cxpb_max_box)
-        
-        adaptive_options_layout.addRow("Crossover Bounds:", crossover_bounds_widget)
-        
-        # Create a widget for mutation bounds
-        mutation_bounds_widget = QWidget()
-        mutation_bounds_layout = QHBoxLayout(mutation_bounds_widget)
-        mutation_bounds_layout.setContentsMargins(0, 0, 0, 0)
-        
-        self.mutpb_min_box = QDoubleSpinBox()
-        self.mutpb_min_box.setRange(0.01, 0.2)
-        self.mutpb_min_box.setValue(0.05)
-        self.mutpb_min_box.setDecimals(2)
-        self.mutpb_min_box.setSingleStep(0.01)
-        self.mutpb_min_box.setToolTip("Minimum mutation probability")
-        
-        self.mutpb_max_box = QDoubleSpinBox()
-        self.mutpb_max_box.setRange(0.2, 0.8)
-        self.mutpb_max_box.setValue(0.5)
-        self.mutpb_max_box.setDecimals(2)
-        self.mutpb_max_box.setSingleStep(0.05)
-        self.mutpb_max_box.setToolTip("Maximum mutation probability")
-        
-        mutation_bounds_layout.addWidget(QLabel("Min:"))
-        mutation_bounds_layout.addWidget(self.mutpb_min_box)
-        mutation_bounds_layout.addWidget(QLabel("Max:"))
-        mutation_bounds_layout.addWidget(self.mutpb_max_box)
-        
-        adaptive_options_layout.addRow("Mutation Bounds:", mutation_bounds_widget)
-        
-        # Initially hide adaptive options
-        self.adaptive_rates_options.setVisible(False)
-
-        ga_hyper_layout.addRow("Population Size:", self.ga_pop_size_box)
-        ga_hyper_layout.addRow("Number of Generations:", self.ga_num_generations_box)
-        ga_hyper_layout.addRow("Crossover Probability (cxpb):", self.ga_cxpb_box)
-        ga_hyper_layout.addRow("Mutation Probability (mutpb):", self.ga_mutpb_box)
-        ga_hyper_layout.addRow("Tolerance (tol):", self.ga_tol_box)
-        ga_hyper_layout.addRow("Sparsity Penalty (alpha):", self.ga_alpha_box)
-        ga_hyper_layout.addRow("Benchmark Runs:", self.ga_benchmark_runs_box)
-        ga_hyper_layout.addRow("", self.adaptive_rates_checkbox)
-        ga_hyper_layout.addRow("", self.adaptive_rates_options)
-
-        # Add a small Run GA button in the hyperparameters sub-tab
-        self.hyper_run_ga_button = QPushButton("Run GA")
-        self.hyper_run_ga_button.setFixedWidth(100)
-        self.hyper_run_ga_button.clicked.connect(self.run_ga)
-        ga_hyper_layout.addRow("Run GA:", self.hyper_run_ga_button)
-
-        # -------------------- Sub-tab 2: DVA Parameters --------------------
-        ga_param_tab = QWidget()
-        ga_param_layout = QVBoxLayout(ga_param_tab)
-
-        self.ga_param_table = QTableWidget()
-        dva_parameters = [
-            *[f"beta_{i}" for i in range(1,16)],
-            *[f"lambda_{i}" for i in range(1,16)],
-            *[f"mu_{i}" for i in range(1,4)],
-            *[f"nu_{i}" for i in range(1,16)]
-        ]
-        self.ga_param_table.setRowCount(len(dva_parameters))
-        self.ga_param_table.setColumnCount(5)
-        self.ga_param_table.setHorizontalHeaderLabels(
-            ["Parameter", "Fixed", "Fixed Value", "Lower Bound", "Upper Bound"]
-        )
-        self.ga_param_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.ga_param_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-
-        for row, param in enumerate(dva_parameters):
-            param_item = QTableWidgetItem(param)
-            param_item.setFlags(Qt.ItemIsEnabled)
-            self.ga_param_table.setItem(row, 0, param_item)
-
-            fixed_checkbox = QCheckBox()
-            fixed_checkbox.setChecked(True)  # Set fixed to true by default
-            fixed_checkbox.stateChanged.connect(lambda state, r=row: self.toggle_ga_fixed(state, r))
-            self.ga_param_table.setCellWidget(row, 1, fixed_checkbox)
-
-            fixed_value_spin = QDoubleSpinBox()
-            fixed_value_spin.setRange(0, 10e9)  # Changed to 0-10e9 range
-            fixed_value_spin.setDecimals(6)
-            fixed_value_spin.setValue(0.0)  # Set fixed value to 0
-            fixed_value_spin.setEnabled(True)  # Enable because fixed is checked
-            self.ga_param_table.setCellWidget(row, 2, fixed_value_spin)
-
-            lower_bound_spin = QDoubleSpinBox()
-            lower_bound_spin.setRange(0, 10e9)  # Changed to 0-10e9 range
-            lower_bound_spin.setDecimals(6)
-            lower_bound_spin.setValue(0.0)  # Set to 0
-            lower_bound_spin.setEnabled(False)  # Disable because fixed is checked
-            self.ga_param_table.setCellWidget(row, 3, lower_bound_spin)
-
-            upper_bound_spin = QDoubleSpinBox()
-            upper_bound_spin.setRange(0, 10e9)  # Changed to 0-10e9 range
-            upper_bound_spin.setDecimals(6)
-            upper_bound_spin.setValue(1.0)  # Set to 1
-            upper_bound_spin.setEnabled(False)  # Disable because fixed is checked
-            self.ga_param_table.setCellWidget(row, 4, upper_bound_spin)
-
-        ga_param_layout.addWidget(self.ga_param_table)
-
-        # -------------------- Sub-tab 3: Results --------------------
-        ga_results_tab = QWidget()
-        ga_results_layout = QVBoxLayout(ga_results_tab)
-
-        # Create a header area for label and export button
-        header_container = QWidget()
-        header_layout = QHBoxLayout(header_container)
-        header_layout.setContentsMargins(0, 0, 0, 0) # No margins for this internal layout
-
-        results_label = QLabel("GA Optimization Results:")
-        header_layout.addWidget(results_label)
-        header_layout.addStretch() # Add spacer to push the export button to the right
-
-        self.export_ga_results_button = QPushButton("Export GA Results")
-        self.export_ga_results_button.setObjectName("secondary-button") # Use existing styling if desired
-        self.export_ga_results_button.setToolTip("Export the GA optimization results to a JSON file")
-        self.export_ga_results_button.setEnabled(False)  # Initially disabled
-        # self.export_ga_results_button.clicked.connect(self.export_ga_results_to_file) # Will connect this later
-        header_layout.addWidget(self.export_ga_results_button)
-        
-        ga_results_layout.addWidget(header_container) # Add the header with label and button
-        
-        self.ga_results_text = QTextEdit()
-        self.ga_results_text.setReadOnly(True)
-        ga_results_layout.addWidget(self.ga_results_text)
-
-        # -------------------- Sub-tab 4: Benchmarking --------------------
-        ga_benchmark_tab = QWidget()
-        ga_benchmark_layout = QVBoxLayout(ga_benchmark_tab)
-
-        # Create buttons for import/export
-        button_container = QWidget()
-        button_layout = QHBoxLayout(button_container)
-        button_layout.setContentsMargins(0, 0, 0, 10)  # Add some bottom margin
-
-        self.import_benchmark_button = QPushButton("Import Benchmark Data")
-        self.import_benchmark_button.setToolTip("Import previously saved GA benchmark data")
-        self.import_benchmark_button.clicked.connect(self.import_ga_benchmark_data)
-        button_layout.addWidget(self.import_benchmark_button)
-
-        self.export_benchmark_button = QPushButton("Export Benchmark Data")
-        self.export_benchmark_button.setToolTip("Export current GA benchmark data to a file")
-        self.export_benchmark_button.setEnabled(False)  # Initially disabled until data is available
-        self.export_benchmark_button.clicked.connect(self.export_ga_benchmark_data)
-        button_layout.addWidget(self.export_benchmark_button)
-
-        button_layout.addStretch()  # Add stretch to push buttons to the left
-        ga_benchmark_layout.addWidget(button_container)
-
-        # Create tabs for different benchmark visualizations
-        self.benchmark_viz_tabs = QTabWidget()
-        
-        # Create tabs for different visualizations
-        violin_tab = QWidget()
-        violin_layout = QVBoxLayout(violin_tab)
-        self.violin_plot_widget = QWidget()
-        violin_layout.addWidget(self.violin_plot_widget)
-        
-        dist_tab = QWidget()
-        dist_layout = QVBoxLayout(dist_tab)
-        self.dist_plot_widget = QWidget()
-        dist_layout.addWidget(self.dist_plot_widget)
-        
-        scatter_tab = QWidget()
-        scatter_layout = QVBoxLayout(scatter_tab)
-        self.scatter_plot_widget = QWidget()
-        scatter_layout.addWidget(self.scatter_plot_widget)
-        
-        heatmap_tab = QWidget()
-        heatmap_layout = QVBoxLayout(heatmap_tab)
-        self.heatmap_plot_widget = QWidget()
-        heatmap_layout.addWidget(self.heatmap_plot_widget)
-        
-        # Add Q-Q plot tab
-        qq_tab = QWidget()
-        qq_layout = QVBoxLayout(qq_tab)
-        self.qq_plot_widget = QWidget()
-        qq_layout.addWidget(self.qq_plot_widget)
-        
-        # Summary statistics tabs (create subtabs for better organization)
-        stats_tab = QWidget()
-        stats_layout = QVBoxLayout(stats_tab)
-        
-        # Create a tabbed widget for the statistics section
-        stats_subtabs = QTabWidget()
-        
-        # ---- Subtab 1: Summary Statistics ----
-        summary_tab = QWidget()
-        summary_layout = QVBoxLayout(summary_tab)
-        
-        # Add summary statistics table
-        self.benchmark_stats_table = QTableWidget()
-        self.benchmark_stats_table.setColumnCount(5)
-        self.benchmark_stats_table.setHorizontalHeaderLabels(["Metric", "Min", "Max", "Mean", "Std"])
-        self.benchmark_stats_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        summary_layout.addWidget(QLabel("Statistical Summary of All Runs:"))
-        summary_layout.addWidget(self.benchmark_stats_table)
-        
-        # ---- Subtab 2: All Runs Table ----
-        runs_tab = QWidget()
-        runs_layout = QVBoxLayout(runs_tab)
-        
-        # Create a table for all runs
-        self.benchmark_runs_table = QTableWidget()
-        self.benchmark_runs_table.setColumnCount(4)
-        self.benchmark_runs_table.setHorizontalHeaderLabels(["Run #", "Fitness", "Rank", "Details"])
-        self.benchmark_runs_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.benchmark_runs_table.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.benchmark_runs_table.setSelectionMode(QAbstractItemView.SingleSelection)
-        self.benchmark_runs_table.itemClicked.connect(self.show_run_details)
-        
-        runs_layout.addWidget(QLabel("All Benchmark Runs:"))
-        runs_layout.addWidget(self.benchmark_runs_table)
-        
-        # Create run details text area
-        details_tab = QWidget()
-        details_layout = QVBoxLayout(details_tab)
-        details_group = QGroupBox("Run Details")
-        group_layout = QVBoxLayout(details_group)
-        self.run_details_text = QTextEdit()
-        self.run_details_text.setReadOnly(True)
-        group_layout.addWidget(self.run_details_text)
-        details_layout.addWidget(details_group)
-        
-        # Create GA Operations tab as a subtab
-        ga_ops_tab = QWidget()
-        ga_ops_layout = QVBoxLayout(ga_ops_tab)
-        self.ga_ops_plot_widget = QWidget()
-        ga_ops_layout.addWidget(self.ga_ops_plot_widget)
-        
-        # Add the subtabs to the stats tabbed widget
-        stats_subtabs.addTab(summary_tab, "Summary Statistics")
-        stats_subtabs.addTab(runs_tab, "All Runs")
-        stats_subtabs.addTab(details_tab, "Run Details")
-        stats_subtabs.addTab(ga_ops_tab, "GA Operations")
-        
-        # Add the stats tabbed widget to the stats tab
-        stats_layout.addWidget(stats_subtabs)
-        
-        # Add all visualization tabs to the benchmark visualization tabs
-        self.benchmark_viz_tabs.addTab(violin_tab, "Violin Plot")
-        self.benchmark_viz_tabs.addTab(dist_tab, "Distribution")
-        self.benchmark_viz_tabs.addTab(scatter_tab, "Scatter Plot")
-        self.benchmark_viz_tabs.addTab(heatmap_tab, "Parameter Correlations")
-        self.benchmark_viz_tabs.addTab(qq_tab, "Q-Q Plot")
-        self.benchmark_viz_tabs.addTab(stats_tab, "Statistics")
-        
-        # GA Operations Performance Tab - already added as a subtab of Statistics
-        
-        # Add the benchmark visualization tabs to the benchmark tab
-        ga_benchmark_layout.addWidget(self.benchmark_viz_tabs)
-        
-        # Add all sub-tabs to the GA tab widget
-        # Initialize empty benchmark data storage
-        self.ga_benchmark_data = []
-
-        # Add all sub-tabs to the GA tab widget
-        self.ga_sub_tabs.addTab(ga_hyper_tab, "GA Settings")
-        self.ga_sub_tabs.addTab(ga_param_tab, "DVA Parameters")
-        self.ga_sub_tabs.addTab(ga_results_tab, "Results")
-        self.ga_sub_tabs.addTab(ga_benchmark_tab, "GA Benchmarking")
-
-        # Add the GA sub-tabs widget to the main GA tab layout
-        layout.addWidget(self.ga_sub_tabs)
-        self.ga_tab.setLayout(layout)
-        
-    def toggle_fixed(self, state, row, table=None):
-        """Toggle the fixed state of a DVA parameter row"""
-        if table is None:
-            table = self.dva_param_table
-            
-        fixed = (state == Qt.Checked)
-        fixed_value_spin = table.cellWidget(row, 2)
-        lower_bound_spin = table.cellWidget(row, 3)
-        upper_bound_spin = table.cellWidget(row, 4)
-
-        fixed_value_spin.setEnabled(fixed)
-        lower_bound_spin.setEnabled(not fixed)
-        upper_bound_spin.setEnabled(not fixed)
-
-    def toggle_ga_fixed(self, state, row, table=None):
-        """Toggle the fixed state of a GA parameter row"""
-        if table is None:
-            table = self.ga_param_table
-            
-        fixed = (state == Qt.Checked)
-        fixed_value_spin = table.cellWidget(row, 2)
-        lower_bound_spin = table.cellWidget(row, 3)
-        upper_bound_spin = table.cellWidget(row, 4)
-        
-        # Enable/disable appropriate spinboxes
-        fixed_value_spin.setEnabled(fixed)
-        lower_bound_spin.setEnabled(not fixed)
-        upper_bound_spin.setEnabled(not fixed)
-        
-        # If switching to fixed mode, copy current lower bound value to fixed value
-        if fixed:
-            fixed_value_spin.setValue(lower_bound_spin.value())
-        # If switching to range mode, ensure lower bound is not greater than upper bound
-        else:
-            if lower_bound_spin.value() > upper_bound_spin.value():
-                upper_bound_spin.setValue(lower_bound_spin.value())
-
-    def toggle_adaptive_rates_options(self, state):
-        """Show or hide adaptive rates options based on checkbox state"""
-        self.adaptive_rates_options.setVisible(state == Qt.Checked)
-        
-        # Enable/disable the fixed rate inputs based on adaptive rates setting
-        self.ga_cxpb_box.setEnabled(state != Qt.Checked)
-        self.ga_mutpb_box.setEnabled(state != Qt.Checked)
-        
-        # Update tooltips to indicate that rates will be adaptive
-        if state == Qt.Checked:
-            self.ga_cxpb_box.setToolTip("Starting crossover probability (will adapt during optimization)")
-            self.ga_mutpb_box.setToolTip("Starting mutation probability (will adapt during optimization)")
-        else:
-            self.ga_cxpb_box.setToolTip("Crossover probability")
-            self.ga_mutpb_box.setToolTip("Mutation probability")
 
     def _run_sobol_implementation(self):
         """Run Sobol sensitivity analysis - main implementation"""
@@ -6620,2460 +6215,6 @@ class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin
         # Reset logic for all parameters would be implemented here
 
 
-    def run_ga(self):
-        """Run genetic algorithm optimization"""
-        # Check if a GA worker is already running
-        if hasattr(self, 'ga_worker') and self.ga_worker.isRunning():
-            QMessageBox.warning(self, "Process Running", 
-                               "A Genetic Algorithm optimization is already running. Please wait for it to complete.")
-            return
-            
-        if self.omega_start_box.value() >= self.omega_end_box.value():
-            QMessageBox.warning(self, "Input Error", "Ω Start must be less than Ω End.")
-            return
-
-        target_values, weights = self.get_target_values_weights()
-        
-        # Get GA hyperparameters
-        pop_size = self.ga_pop_size_box.value()
-        num_gen = self.ga_num_generations_box.value()
-        crossover_prob = self.ga_cxpb_box.value()
-        mutation_prob = self.ga_mutpb_box.value()
-        tolerance = self.ga_tol_box.value()
-        alpha = self.ga_alpha_box.value()
-        
-        # Get number of benchmark runs
-        self.benchmark_runs = self.ga_benchmark_runs_box.value()
-        self.current_benchmark_run = 0
-        
-        # Clear benchmark data if running multiple times
-        if self.benchmark_runs > 1:
-            self.ga_benchmark_data = []
-            # Enable the benchmark tab if running multiple times
-            self.ga_sub_tabs.setTabEnabled(self.ga_sub_tabs.indexOf(self.ga_sub_tabs.findChild(QWidget, "GA Benchmarking")), True)
-        
-        # Get DVA parameter bounds
-        dva_bounds = {}
-        EPSILON = 1e-6
-        for row in range(self.ga_param_table.rowCount()):
-            param_item = self.ga_param_table.item(row, 0)
-            param_name = param_item.text()
-            
-            fixed_widget = self.ga_param_table.cellWidget(row, 1)
-            fixed = fixed_widget.isChecked()
-            
-            if fixed:
-                fixed_value_widget = self.ga_param_table.cellWidget(row, 2)
-                fixed_value = fixed_value_widget.value()
-                dva_bounds[param_name] = (fixed_value, fixed_value + EPSILON)
-            else:
-                lower_bound_widget = self.ga_param_table.cellWidget(row, 3)
-                upper_bound_widget = self.ga_param_table.cellWidget(row, 4)
-                lower = lower_bound_widget.value()
-                upper = upper_bound_widget.value()
-                if lower > upper:
-                    QMessageBox.warning(self, "Input Error", 
-                                       f"For parameter {param_name}, lower bound is greater than upper bound.")
-                    return
-                dva_bounds[param_name] = (lower, upper)
-        
-        # Get main system parameters
-        main_params = (
-            self.mu_box.value(),
-            *[b.value() for b in self.landa_boxes],
-            *[b.value() for b in self.nu_boxes],
-            self.a_low_box.value(),
-            self.a_up_box.value(),
-            self.f_1_box.value(),
-            self.f_2_box.value(),
-            self.omega_dc_box.value(),
-            self.zeta_dc_box.value()
-        )
-        
-        # Update button reference to match the actual button in the UI
-        self.run_ga_button = self.hyper_run_ga_button
-        
-        # Disable run buttons during optimization
-        self.run_frf_button.setEnabled(False)
-        self.run_sobol_button.setEnabled(False)
-        self.run_ga_button.setEnabled(False)
-        
-        # Create progress bar if it doesn't exist
-        if not hasattr(self, 'ga_progress_bar'):
-            self.ga_progress_bar = QProgressBar()
-            self.ga_progress_bar.setRange(0, 100)
-            self.ga_progress_bar.setValue(0)
-            self.ga_progress_bar.setTextVisible(True)
-            self.ga_progress_bar.setFormat("GA Progress: %p%")
-            
-            # Find where to add progress bar in the layout
-            ga_results_tab_layout = self.ga_results_text.parent().layout()
-            ga_results_tab_layout.insertWidget(0, self.ga_progress_bar)
-        else:
-            self.ga_progress_bar.setValue(0)
-            
-        # Make sure the progress bar is visible
-        self.ga_progress_bar.show()
-        
-        # Update status
-        self.status_bar.showMessage("Running GA optimization...")
-        self.ga_results_text.append("\n--- Running Genetic Algorithm Optimization ---\n")
-        self.ga_results_text.append(f"Population Size: {pop_size}")
-        self.ga_results_text.append(f"Number of Generations: {num_gen}")
-        self.ga_results_text.append(f"Crossover Probability: {crossover_prob}")
-        self.ga_results_text.append(f"Mutation Probability: {mutation_prob}")
-        self.ga_results_text.append(f"Tolerance: {tolerance}")
-        self.ga_results_text.append(f"Sparsity Penalty (alpha): {alpha}")
-        
-        # Add debug output for adaptive rates
-        adaptive_rates = self.adaptive_rates_checkbox.isChecked()
-        self.ga_results_text.append(f"Adaptive Rates: {'Enabled' if adaptive_rates else 'Disabled'}")
-        if adaptive_rates:
-            self.ga_results_text.append(f"  - Stagnation Limit: {self.stagnation_limit_box.value()}")
-            self.ga_results_text.append(f"  - Crossover Range: {self.cxpb_min_box.value():.2f} - {self.cxpb_max_box.value():.2f}")
-            self.ga_results_text.append(f"  - Mutation Range: {self.mutpb_min_box.value():.2f} - {self.mutpb_max_box.value():.2f}")
-        self.ga_results_text.append("\nStarting optimization...\n")
-        
-        # Create and start worker
-        original_dva_parameter_order = [
-            'beta_1','beta_2','beta_3','beta_4','beta_5','beta_6',
-            'beta_7','beta_8','beta_9','beta_10','beta_11','beta_12',
-            'beta_13','beta_14','beta_15',
-            'lambda_1','lambda_2','lambda_3','lambda_4','lambda_5',
-            'lambda_6','lambda_7','lambda_8','lambda_9','lambda_10',
-            'lambda_11','lambda_12','lambda_13','lambda_14','lambda_15',
-            'mu_1','mu_2','mu_3',
-            'nu_1','nu_2','nu_3','nu_4','nu_5','nu_6',
-            'nu_7','nu_8','nu_9','nu_10','nu_11','nu_12',
-            'nu_13','nu_14','nu_15'
-        ]
-        
-        # Convert dva_bounds and dva_order into ga_parameter_data format
-        ga_parameter_data = []
-        for param_name in original_dva_parameter_order:
-            if param_name in dva_bounds:
-                low, high = dva_bounds[param_name]
-                # Check if parameter is fixed (low == high)
-                fixed = abs(low - high) < EPSILON
-                ga_parameter_data.append((param_name, low, high, fixed))
-                
-        # If there's an existing worker, make sure it's properly cleaned up
-        if hasattr(self, 'ga_worker'):
-            try:
-                self.ga_worker.finished.disconnect()
-                self.ga_worker.error.disconnect()
-                self.ga_worker.update.disconnect()
-                self.ga_worker.progress.disconnect()
-            except Exception:
-                pass
-                
-        # Create a new worker
-        self.ga_worker = GAWorker(
-            main_params=main_params,
-            target_values_dict=target_values,
-            weights_dict=weights,
-            omega_start=self.omega_start_box.value(),
-            omega_end=self.omega_end_box.value(),
-            omega_points=self.omega_points_box.value(),
-            ga_pop_size=pop_size,
-            ga_num_generations=num_gen,
-            ga_cxpb=crossover_prob,
-            ga_mutpb=mutation_prob,
-            ga_tol=tolerance,
-            ga_parameter_data=ga_parameter_data,
-            alpha=alpha,
-            track_metrics=True,  # Enable metrics tracking for visualization
-            adaptive_rates=self.adaptive_rates_checkbox.isChecked(),  # Pass the adaptive rates setting
-            stagnation_limit=self.stagnation_limit_box.value(),  # Get stagnation limit from UI
-            cxpb_min=self.cxpb_min_box.value(),  # Get min crossover probability
-            cxpb_max=self.cxpb_max_box.value(),  # Get max crossover probability
-            mutpb_min=self.mutpb_min_box.value(),  # Get min mutation probability
-            mutpb_max=self.mutpb_max_box.value()  # Get max mutation probability
-        )
-        
-        # Connect signals using strong references to avoid premature garbage collection
-        self.ga_worker.finished.connect(self.handle_ga_finished)
-        self.ga_worker.error.connect(self.handle_ga_error)
-        self.ga_worker.update.connect(self.handle_ga_update)
-        self.ga_worker.progress.connect(self.update_ga_progress)
-        
-        # Set up a watchdog timer for the GA worker
-        if hasattr(self, 'ga_watchdog_timer'):
-            self.ga_watchdog_timer.stop()
-        else:
-            self.ga_watchdog_timer = QTimer(self)
-            self.ga_watchdog_timer.timeout.connect(self.check_ga_worker_health)
-            
-        self.ga_watchdog_timer.start(10000)  # Check every 10 seconds
-        
-        # Start the worker
-        self.ga_worker.start()
-        
-    def check_ga_worker_health(self):
-        """Check if the GA worker is still responsive"""
-        if hasattr(self, 'ga_worker') and self.ga_worker.isRunning():
-            # The worker is still running, which is good
-            # We could add more sophisticated checks here if needed
-            pass
-        else:
-            # The worker is not running anymore, stop the watchdog
-            if hasattr(self, 'ga_watchdog_timer'):
-                self.ga_watchdog_timer.stop()
-                
-    def update_ga_progress(self, value):
-        """Update the GA progress bar, accounting for multiple benchmark runs"""
-        if hasattr(self, 'ga_progress_bar'):
-            if hasattr(self, 'benchmark_runs') and self.benchmark_runs > 1:
-                # Calculate overall progress across all runs
-                # Each run contributes (1/total_runs) of the progress
-                run_contribution = 100.0 / self.benchmark_runs
-                current_run_progress = value / 100.0  # Convert to fraction
-                # Add progress from completed runs plus fractional progress from current run
-                overall_progress = ((self.current_benchmark_run - 1) * run_contribution) + (current_run_progress * run_contribution)
-                self.ga_progress_bar.setValue(int(overall_progress))
-            else:
-                # Single run - direct progress
-                self.ga_progress_bar.setValue(value)
-            
-    def handle_ga_finished(self, results, best_ind, parameter_names, best_fitness):
-        """Handle the completion of the GA optimization"""
-        # Stop the watchdog timer
-        if hasattr(self, 'ga_watchdog_timer'):
-            self.ga_watchdog_timer.stop()
-        
-        # For benchmarking, collect data from this run
-        self.current_benchmark_run += 1
-        
-        # Store benchmark results
-        if hasattr(self, 'benchmark_runs') and self.benchmark_runs > 1:
-            # Create a data dictionary for this run
-            run_data = {
-                'run_number': self.current_benchmark_run,
-                'best_fitness': best_fitness,
-                'best_solution': list(best_ind),
-                'parameter_names': parameter_names
-            }
-            
-            # Add any additional metrics from results
-            if isinstance(results, dict):
-                for key, value in results.items():
-                    if isinstance(value, (int, float)) and np.isfinite(value):
-                        run_data[key] = value
-
-                # Add benchmark metrics if available
-                if 'benchmark_metrics' in results:
-                    run_data['benchmark_metrics'] = results['benchmark_metrics']
-            
-            # Store the run data
-            self.ga_benchmark_data.append(run_data)
-            
-            # Update the status message
-            self.status_bar.showMessage(f"GA run {self.current_benchmark_run} of {self.benchmark_runs} completed")
-            
-            # Update progress bar to show completed percentage of all runs
-            if hasattr(self, 'ga_progress_bar'):
-                progress = int(self.current_benchmark_run * 100 / self.benchmark_runs)
-                self.ga_progress_bar.setValue(progress)
-            
-            # Check if we need to run again
-            if self.current_benchmark_run < self.benchmark_runs:
-                self.ga_results_text.append(f"\n--- Run {self.current_benchmark_run} completed, starting run {self.current_benchmark_run + 1}/{self.benchmark_runs} ---")
-                # Set up for next run
-                QTimer.singleShot(100, self.run_next_ga_benchmark)
-                return
-            else:
-                # All runs completed, visualize the benchmark results
-                self.visualize_ga_benchmark_results()
-                self.export_benchmark_button.setEnabled(True)
-                self.ga_results_text.append(f"\n--- All {self.benchmark_runs} benchmark runs completed ---")
-        else:
-            # For single runs, store the data directly
-            run_data = {
-                'run_number': 1,
-                'best_fitness': best_fitness,
-                'best_solution': list(best_ind),
-                'parameter_names': parameter_names
-            }
-            
-            # Add benchmark metrics if available
-            if isinstance(results, dict) and 'benchmark_metrics' in results:
-                run_data['benchmark_metrics'] = results['benchmark_metrics']
-            
-            self.ga_benchmark_data = [run_data]
-            self.visualize_ga_benchmark_results()
-                
-        # Re-enable buttons when completely done
-        self.run_frf_button.setEnabled(True)
-        self.run_sobol_button.setEnabled(True)
-        self.run_ga_button.setEnabled(True)
-        
-        self.status_bar.showMessage("GA optimization completed")
-        
-        # Only show detailed results for single runs or the final benchmark run
-        if not hasattr(self, 'benchmark_runs') or self.benchmark_runs == 1 or self.current_benchmark_run == self.benchmark_runs:
-            self.ga_results_text.append("\n--- GA Optimization Completed ---")
-            self.ga_results_text.append(f"Best fitness: {best_fitness:.6f}")
-        self.ga_results_text.append("\nBest Parameters:")
-        
-        # Check if there are any warnings in the results
-        if isinstance(results, dict) and "Warning" in results:
-            self.ga_results_text.append(f"\nWarning: {results['Warning']}")
-        
-        # Create a dictionary mapping parameter names to their values
-        best_params = {name: value for name, value in zip(parameter_names, best_ind)}
-        
-        # Store best parameters for easy access later
-        self.current_ga_best_params = best_params
-        self.current_ga_best_fitness = best_fitness
-        self.current_ga_full_results = results
-        
-        for param_name, value in best_params.items():
-            self.ga_results_text.append(f"  {param_name}: {value:.6f}")
-            
-        # If we have actual results, show them
-        if isinstance(results, dict) and "singular_response" in results:
-            self.ga_results_text.append(f"\nFinal Singular Response: {results['singular_response']:.6f}")
-        
-    def handle_ga_error(self, error_msg):
-        """Handle errors from the GA worker"""
-        # Stop the watchdog timer
-        if hasattr(self, 'ga_watchdog_timer'):
-            self.ga_watchdog_timer.stop()
-            
-        # Hide or reset the progress bar
-        if hasattr(self, 'ga_progress_bar'):
-            self.ga_progress_bar.setValue(0)
-            
-        QMessageBox.critical(self, "Error in GA Optimization", str(error_msg))
-        self.status_bar.showMessage("GA optimization failed")
-        self.ga_results_text.append(f"\nError in GA optimization: {error_msg}")
-        
-        # Make sure to re-enable buttons
-        self.run_frf_button.setEnabled(True)
-        self.run_sobol_button.setEnabled(True)
-        self.run_ga_button.setEnabled(True)
-        
-        # Try to recover by cleaning up any residual state
-        if hasattr(self, 'ga_worker'):
-            try:
-                # Attempt to terminate the worker if it's still running
-                if self.ga_worker.isRunning():
-                    self.ga_worker.terminate()
-                    self.ga_worker.wait(1000)  # Wait up to 1 second for it to finish
-            except Exception as e:
-                print(f"Error cleaning up GA worker: {str(e)}")
-        
-    def handle_ga_update(self, msg):
-        """Handle update messages from the GA worker"""
-        self.ga_results_text.append(msg)
-        # Auto-scroll to the bottom to show latest messages
-        self.ga_results_text.verticalScrollBar().setValue(
-            self.ga_results_text.verticalScrollBar().maximum()
-        )
-        
-    def run_next_ga_benchmark(self):
-        """Run the next GA benchmark iteration"""
-        # Clear the existing GA worker to start fresh
-        if hasattr(self, 'ga_worker'):
-            try:
-                self.ga_worker.finished.disconnect()
-                self.ga_worker.error.disconnect()
-                self.ga_worker.update.disconnect()
-                self.ga_worker.progress.disconnect()
-            except Exception:
-                pass
-        
-        # Get the required parameters again
-        target_values, weights = self.get_target_values_weights()
-        
-        # Get GA hyperparameters
-        pop_size = self.ga_pop_size_box.value()
-        num_gen = self.ga_num_generations_box.value()
-        crossover_prob = self.ga_cxpb_box.value()
-        mutation_prob = self.ga_mutpb_box.value()
-        tolerance = self.ga_tol_box.value()
-        alpha = self.ga_alpha_box.value()
-        
-        # Get DVA parameter bounds
-        dva_bounds = {}
-        EPSILON = 1e-6
-        for row in range(self.ga_param_table.rowCount()):
-            param_item = self.ga_param_table.item(row, 0)
-            param_name = param_item.text()
-            
-            fixed_widget = self.ga_param_table.cellWidget(row, 1)
-            fixed = fixed_widget.isChecked()
-            
-            if fixed:
-                fixed_value_widget = self.ga_param_table.cellWidget(row, 2)
-                fixed_value = fixed_value_widget.value()
-                dva_bounds[param_name] = (fixed_value, fixed_value + EPSILON)
-            else:
-                lower_bound_widget = self.ga_param_table.cellWidget(row, 3)
-                upper_bound_widget = self.ga_param_table.cellWidget(row, 4)
-                lower = lower_bound_widget.value()
-                upper = upper_bound_widget.value()
-                dva_bounds[param_name] = (lower, upper)
-        
-        # Get main system parameters
-        main_params = (
-            self.mu_box.value(),
-            *[b.value() for b in self.landa_boxes],
-            *[b.value() for b in self.nu_boxes],
-            self.a_low_box.value(),
-            self.a_up_box.value(),
-            self.f_1_box.value(),
-            self.f_2_box.value(),
-            self.omega_dc_box.value(),
-            self.zeta_dc_box.value()
-        )
-        
-        # Reset progress bar
-        if hasattr(self, 'ga_progress_bar'):
-            self.ga_progress_bar.setValue(0)
-            
-        # Make sure the progress bar is visible
-        self.ga_progress_bar.show()
-        
-        # Update status
-        self.status_bar.showMessage(f"Running GA optimization (Run {self.current_benchmark_run + 1}/{self.benchmark_runs})...")
-        
-        # Create and start worker
-        original_dva_parameter_order = [
-            'beta_1','beta_2','beta_3','beta_4','beta_5','beta_6',
-            'beta_7','beta_8','beta_9','beta_10','beta_11','beta_12',
-            'beta_13','beta_14','beta_15',
-            'lambda_1','lambda_2','lambda_3','lambda_4','lambda_5',
-            'lambda_6','lambda_7','lambda_8','lambda_9','lambda_10',
-            'lambda_11','lambda_12','lambda_13','lambda_14','lambda_15',
-            'mu_1','mu_2','mu_3',
-            'nu_1','nu_2','nu_3','nu_4','nu_5','nu_6',
-            'nu_7','nu_8','nu_9','nu_10','nu_11','nu_12',
-            'nu_13','nu_14','nu_15'
-        ]
-        
-        # Convert dva_bounds and dva_order into ga_parameter_data format
-        ga_parameter_data = []
-        for param_name in original_dva_parameter_order:
-            if param_name in dva_bounds:
-                low, high = dva_bounds[param_name]
-                # Check if parameter is fixed (low == high)
-                fixed = abs(low - high) < EPSILON
-                ga_parameter_data.append((param_name, low, high, fixed))
-        
-        # Create a new worker
-        self.ga_worker = GAWorker(
-            main_params=main_params,
-            target_values_dict=target_values,
-            weights_dict=weights,
-            omega_start=self.omega_start_box.value(),
-            omega_end=self.omega_end_box.value(),
-            omega_points=self.omega_points_box.value(),
-            ga_pop_size=pop_size,
-            ga_num_generations=num_gen,
-            ga_cxpb=crossover_prob,
-            ga_mutpb=mutation_prob,
-            ga_tol=tolerance,
-            ga_parameter_data=ga_parameter_data,
-            alpha=alpha,
-            track_metrics=True,  # Enable metrics tracking for visualization
-            adaptive_rates=self.adaptive_rates_checkbox.isChecked(),  # Pass the adaptive rates setting
-            stagnation_limit=self.stagnation_limit_box.value(),  # Get stagnation limit from UI
-            cxpb_min=self.cxpb_min_box.value(),  # Get min crossover probability
-            cxpb_max=self.cxpb_max_box.value(),  # Get max crossover probability
-            mutpb_min=self.mutpb_min_box.value(),  # Get min mutation probability
-            mutpb_max=self.mutpb_max_box.value()  # Get max mutation probability
-        )
-        
-        # Connect signals using strong references to avoid premature garbage collection
-        self.ga_worker.finished.connect(self.handle_ga_finished)
-        self.ga_worker.error.connect(self.handle_ga_error)
-        self.ga_worker.update.connect(self.handle_ga_update)
-        self.ga_worker.progress.connect(self.update_ga_progress)
-        
-        # Set up a watchdog timer for the GA worker
-        if hasattr(self, 'ga_watchdog_timer'):
-            self.ga_watchdog_timer.stop()
-        else:
-            self.ga_watchdog_timer = QTimer(self)
-            self.ga_watchdog_timer.timeout.connect(self.check_ga_worker_health)
-            
-        self.ga_watchdog_timer.start(10000)  # Check every 10 seconds
-        
-        # Start the worker
-        self.ga_worker.start()
-    
-    def _open_plot_window(self, fig, title):
-        """Opens a new window to display a matplotlib figure."""
-        plot_window = PlotWindow(fig, title)
-        plot_window.setMinimumSize(800, 600)
-        plot_window.show()
-        # Keep a reference to prevent garbage collection
-        if not hasattr(self, '_plot_windows'):
-            self._plot_windows = []
-        self._plot_windows.append(plot_window)
-    
-    def visualize_ga_benchmark_results(self):
-        """Create visualizations for GA benchmark results"""
-        if not hasattr(self, 'ga_benchmark_data') or not self.ga_benchmark_data:
-            return
-            
-        import matplotlib.pyplot as plt
-        import numpy as np
-        import pandas as pd
-        from matplotlib.figure import Figure
-        from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-        import seaborn as sns
-        from computational_metrics_new import visualize_all_metrics
-        
-        # Convert benchmark data to DataFrame for easier analysis
-        df = pd.DataFrame(self.ga_benchmark_data)
-        
-        # Visualize computational metrics
-        widgets_dict = {
-            'ga_ops_plot_widget': self.ga_ops_plot_widget
-        }
-        visualize_all_metrics(widgets_dict, df)
-        
-        # 1. Create violin & box plot
-        try:
-            # Clear existing plot layout
-            if self.violin_plot_widget.layout():
-                for i in reversed(range(self.violin_plot_widget.layout().count())): 
-                    self.violin_plot_widget.layout().itemAt(i).widget().setParent(None)
-            else:
-                self.violin_plot_widget.setLayout(QVBoxLayout())
-                
-            # Create figure for violin/box plot
-            fig_violin = Figure(figsize=(10, 6), tight_layout=True)
-            ax_violin = fig_violin.add_subplot(111)
-            
-            # Create violin plot with box plot inside
-            violin = sns.violinplot(y=df["best_fitness"], ax=ax_violin, inner="box", color="skyblue", orient="v")
-            ax_violin.set_title("Distribution of Best Fitness Values", fontsize=14)
-            ax_violin.set_ylabel("Fitness Value", fontsize=12)
-            ax_violin.grid(True, linestyle="--", alpha=0.7)
-            
-            # Add statistical annotations
-            mean_fitness = df["best_fitness"].mean()
-            median_fitness = df["best_fitness"].median()
-            min_fitness = df["best_fitness"].min()
-            max_fitness = df["best_fitness"].max()
-            std_fitness = df["best_fitness"].std()
-            
-            # Get tolerance value
-            tolerance = self.ga_tol_box.value()
-            
-            # Calculate additional statistics
-            q1 = df["best_fitness"].quantile(0.25)
-            q3 = df["best_fitness"].quantile(0.75)
-            iqr = q3 - q1
-            below_tolerance_count = len(df[df["best_fitness"] <= tolerance])
-            below_tolerance_percent = (below_tolerance_count / len(df)) * 100
-            
-            # Create a legend with enhanced statistical information
-            legend_col1_text = (
-                f"Mean: {mean_fitness:.6f}\n"
-                f"Median: {median_fitness:.6f}\n"
-                f"Min: {min_fitness:.6f}\n"
-                f"Max: {max_fitness:.6f}\n"
-                f"Std Dev: {std_fitness:.6f}"
-            )
-
-            legend_col2_text = (
-                f"Q1 (25%): {q1:.6f}\n"
-                f"Q3 (75%): {q3:.6f}\n"
-                f"IQR: {iqr:.6f}\n"
-                f"Tolerance: {tolerance:.6f}\n"
-                f"Below Tolerance: {below_tolerance_count}/{len(df)} ({below_tolerance_percent:.1f}%)\n"
-                f"Total Runs: {len(df)}"
-            )
-            
-            # Create two text boxes for the legend
-            props = dict(boxstyle='round', facecolor='wheat', alpha=0.5) # Adjusted alpha
-            ax_violin.text(0.05, 0.95, legend_col1_text, transform=ax_violin.transAxes, 
-                    fontsize=12, verticalalignment='top', bbox=props) # Adjusted fontsize
-            ax_violin.text(0.28, 0.95, legend_col2_text, transform=ax_violin.transAxes, 
-                    fontsize=12, verticalalignment='top', bbox=props) # Adjusted fontsize and position
-                    
-            # Add percentile lines with labels (without redundant legend entries)
-            percentiles = [25, 50, 75]
-            percentile_values = df["best_fitness"].quantile(np.array(percentiles) / 100)
-            
-            # Add horizontal lines for percentiles
-            for percentile, value in zip(percentiles, percentile_values):
-                if percentile == 25:
-                    color = 'orange'
-                    linestyle = '--'
-                elif percentile == 50:
-                    color = 'red'
-                    linestyle = '-'
-                elif percentile == 75:
-                    color = 'green'
-                    linestyle = ':'
-                else:
-                    color = 'gray'
-                    linestyle = '-'
-
-                ax_violin.axhline(y=value, color=color, 
-                                 linestyle=linestyle, 
-                                 alpha=0.7, 
-                                 label=f'{percentile}th Percentile')
-            
-            # Add mean and median lines
-            ax_violin.axhline(y=mean_fitness, color='blue', linestyle='-', linewidth=1.5, alpha=0.8, label='Mean')
-            ax_violin.axhline(y=median_fitness, color='purple', linestyle='--', linewidth=1.5, alpha=0.8, label='Median')
-
-            # Add tolerance line with distinct appearance
-            ax_violin.axhline(y=tolerance, color='magenta', linestyle='--', linewidth=2.5, alpha=0.9, 
-                           label=f'Tolerance')
-            
-            # Add a shaded region below tolerance (without redundant legend entry)
-            ax_violin.axhspan(0, tolerance, color='magenta', alpha=0.1, label=None)
-            
-            # Add compact legend for all lines
-            ax_violin.legend(loc='upper right', framealpha=0.7, fontsize=9)
-            
-            # Create canvas and add to layout
-            canvas_violin = FigureCanvasQTAgg(fig_violin)
-            self.violin_plot_widget.layout().addWidget(canvas_violin)
-            
-            # Add toolbar for interactive features
-            from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
-            toolbar_violin = NavigationToolbar(canvas_violin, self.violin_plot_widget)
-            self.violin_plot_widget.layout().addWidget(toolbar_violin)
-
-            # Add "Open in New Window" button
-            open_new_window_button = QPushButton("Open in New Window")
-            open_new_window_button.setObjectName("secondary-button")
-            open_new_window_button.clicked.connect(lambda: self._open_plot_window(fig_violin, "Violin Plot"))
-            self.violin_plot_widget.layout().addWidget(open_new_window_button)
-
-        except Exception as e:
-            print(f"Error creating violin plot: {str(e)}")
-            
-        # 2. Create distribution plots
-        try:
-            # Clear existing plot layout
-            if self.dist_plot_widget.layout():
-                for i in reversed(range(self.dist_plot_widget.layout().count())): 
-                    self.dist_plot_widget.layout().itemAt(i).widget().setParent(None)
-            else:
-                self.dist_plot_widget.setLayout(QVBoxLayout())
-                
-            # Create figure for distribution plot
-            fig_dist = Figure(figsize=(10, 6), tight_layout=True)
-            ax_dist = fig_dist.add_subplot(111)
-            
-            # Create KDE plot with histogram
-            sns.histplot(df["best_fitness"], kde=True, ax=ax_dist, color="skyblue", 
-                        edgecolor="darkblue", alpha=0.5)
-            ax_dist.set_title("Distribution of Best Fitness Values", fontsize=14)
-            ax_dist.set_xlabel("Fitness Value", fontsize=12)
-            ax_dist.set_ylabel("Frequency", fontsize=12)
-            ax_dist.grid(True, linestyle="--", alpha=0.7)
-            
-            # Add vertical line for mean and median (compact legend)
-            mean_fitness = df["best_fitness"].mean()
-            median_fitness = df["best_fitness"].median()
-            std_fitness = df["best_fitness"].std()
-            ax_dist.axvline(mean_fitness, color='red', linestyle='--', linewidth=2, label='Mean')
-            ax_dist.axvline(median_fitness, color='green', linestyle=':', linewidth=2, label='Median')
-            
-            # Add std deviation range (no legend entry)
-            ax_dist.axvspan(mean_fitness - std_fitness, mean_fitness + std_fitness, alpha=0.15, color='yellow', 
-                          label=None)
-            
-            # Add tolerance line
-            tolerance = self.ga_tol_box.value()
-            ax_dist.axvline(tolerance, color='magenta', linestyle='--', linewidth=2.5, alpha=0.9, 
-                          label='Tolerance')
-            
-            # Add a shaded region below tolerance (no legend entry)
-            ax_dist.axvspan(0, tolerance, color='magenta', alpha=0.1, label=None)
-            
-            # Calculate statistics for annotation
-            below_tolerance_count = len(df[df["best_fitness"] <= tolerance])
-            below_tolerance_percent = (below_tolerance_count / len(df)) * 100
-            
-            # Add compact, non-redundant statistics
-            stats_text = (
-                f"Runs: {len(df)}\n"
-                f"Success: {below_tolerance_percent:.1f}%\n"
-                f"Mean: {mean_fitness:.6f}\n"
-                f"Std Dev: {std_fitness:.6f}"
-            )
-            props = dict(boxstyle='round', facecolor='lightblue', alpha=0.6)
-            ax_dist.text(0.95, 0.3, stats_text, transform=ax_dist.transAxes, 
-                      fontsize=10, verticalalignment='top', horizontalalignment='right', bbox=props)
-                      
-            # Add more compact legend
-            ax_dist.legend(loc='upper left', framealpha=0.7, fontsize=9)
-            
-            # Create canvas and add to layout
-            canvas_dist = FigureCanvasQTAgg(fig_dist)
-            self.dist_plot_widget.layout().addWidget(canvas_dist)
-            
-            # Add toolbar for interactive features
-            toolbar_dist = NavigationToolbar(canvas_dist, self.dist_plot_widget)
-            self.dist_plot_widget.layout().addWidget(toolbar_dist)
-
-            # Add "Open in New Window" button
-            open_new_window_button = QPushButton("Open in New Window")
-            open_new_window_button.setObjectName("secondary-button")
-            open_new_window_button.clicked.connect(lambda: self._open_plot_window(fig_dist, "Distribution Plot"))
-            self.dist_plot_widget.layout().addWidget(open_new_window_button)
-
-        except Exception as e:
-            print(f"Error creating distribution plot: {str(e)}")
-            
-        # 3. Create scatter plots
-        try:
-            # Clear existing plot layout
-            if self.scatter_plot_widget.layout():
-                for i in reversed(range(self.scatter_plot_widget.layout().count())): 
-                    self.scatter_plot_widget.layout().itemAt(i).widget().setParent(None)
-            else:
-                self.scatter_plot_widget.setLayout(QVBoxLayout())
-                
-            # Create figure for scatter plot
-            fig_scatter = Figure(figsize=(10, 6), tight_layout=True)
-            ax_scatter = fig_scatter.add_subplot(111)
-            
-            # Create scatter plot of fitness vs run number with trend line
-            from scipy import stats
-            
-            # Calculate linear regression and correlation
-            slope, intercept, r_value, p_value, std_err = stats.linregress(df["run_number"], df["best_fitness"])
-            correlation = r_value
-            
-            # Create scatter plot with trend line
-            sns.regplot(x="run_number", y="best_fitness", data=df, ax=ax_scatter, 
-                       scatter_kws={"color": "darkblue", "alpha": 0.6, "s": 50},
-                       line_kws={"color": "red", "alpha": 0.7})
-            
-            trend_direction = "improving" if slope < 0 else "worsening" if slope > 0 else "stable"
-            ax_scatter.set_title(f"Best Fitness Values Across Runs (Trend: {trend_direction})", fontsize=14)
-            ax_scatter.set_xlabel("Run Number", fontsize=12)
-            ax_scatter.set_ylabel("Best Fitness Value", fontsize=12)
-            ax_scatter.grid(True, linestyle="--", alpha=0.7)
-            
-            # Add tolerance line (without legend entry)
-            tolerance = self.ga_tol_box.value()
-            ax_scatter.axhline(y=tolerance, color='magenta', linestyle='--', linewidth=2.5, alpha=0.9,
-                             label=None)
-            
-            # Add a shaded region below tolerance (no legend entry)
-            ax_scatter.axhspan(0, tolerance, color='magenta', alpha=0.1, label=None)
-            
-            # Color points that are below tolerance
-            below_tolerance_df = df[df["best_fitness"] <= tolerance]
-            below_tolerance_count = len(below_tolerance_df)
-            below_tolerance_percent = (below_tolerance_count / len(df)) * 100
-            
-            if not below_tolerance_df.empty:
-                ax_scatter.scatter(below_tolerance_df["run_number"], below_tolerance_df["best_fitness"], 
-                                 color='green', s=80, alpha=0.8, edgecolor='black', zorder=5,
-                                 label='Success Points')
-            
-            # Find and mark best run
-            best_run_idx = df["best_fitness"].idxmin()
-            best_run = df.iloc[best_run_idx]
-            ax_scatter.scatter(best_run["run_number"], best_run["best_fitness"], 
-                             color='gold', s=120, alpha=1.0, edgecolor='black', marker='*', zorder=6,
-                             label='Best Run')
-            
-            # Add correlation statistics in lower left (away from points)
-            stats_text = (
-                f"Correlation: {correlation:.4f}\n"
-                f"Success Rate: {below_tolerance_percent:.1f}%\n"
-                f"Best: {best_run['best_fitness']:.6f} (Tol: {tolerance:.6f})"
-            )
-            props = dict(boxstyle='round', facecolor='lightyellow', alpha=0.6)
-            ax_scatter.text(0.03, 0.15, stats_text, transform=ax_scatter.transAxes, 
-                         fontsize=10, verticalalignment='bottom', bbox=props)
-            
-            # Add legend with fewer items
-            ax_scatter.legend(loc='lower right', framealpha=0.7)
-            
-            # Create canvas and add to layout
-            canvas_scatter = FigureCanvasQTAgg(fig_scatter)
-            self.scatter_plot_widget.layout().addWidget(canvas_scatter)
-            
-            # Add toolbar for interactive features
-            toolbar_scatter = NavigationToolbar(canvas_scatter, self.scatter_plot_widget)
-            self.scatter_plot_widget.layout().addWidget(toolbar_scatter)
-
-            # Add "Open in New Window" button
-            open_new_window_button = QPushButton("Open in New Window")
-            open_new_window_button.setObjectName("secondary-button")
-            open_new_window_button.clicked.connect(lambda: self._open_plot_window(fig_scatter, "Scatter Plot"))
-            self.scatter_plot_widget.layout().addWidget(open_new_window_button)
-
-        except Exception as e:
-            print(f"Error creating scatter plot: {str(e)}")
-            
-        # 4. Create heatmap of correlation between parameters and fitness
-        try:
-            # Clear existing plot layout
-            if self.heatmap_plot_widget.layout():
-                for i in reversed(range(self.heatmap_plot_widget.layout().count())): 
-                    self.heatmap_plot_widget.layout().itemAt(i).widget().setParent(None)
-            else:
-                self.heatmap_plot_widget.setLayout(QVBoxLayout())
-            
-            # Create figure for heatmap
-            fig_heatmap = Figure(figsize=(12, 10), tight_layout=True)
-            ax_heatmap = fig_heatmap.add_subplot(111)
-            
-            # Extract parameter values from each run into a DataFrame
-            param_values = []
-            
-            if len(df) > 0 and 'best_solution' in df.iloc[0] and 'parameter_names' in df.iloc[0]:
-                # Get parameter names
-                param_names = df.iloc[0]['parameter_names']
-                
-                # Limit to max 10 parameters to keep visualization manageable
-                max_params = min(10, len(param_names))
-                selected_params = param_names[:max_params]
-                
-                # For each run, extract the parameter values
-                for i, row in df.iterrows():
-                    run_data = {'run_number': row['run_number'], 'best_fitness': row['best_fitness']}
-                    
-                    # Extract the parameter values
-                    solution = row['best_solution']
-                    for j, param in enumerate(selected_params):
-                        if j < len(solution):
-                            run_data[param] = solution[j]
-                    
-                    param_values.append(run_data)
-                
-                # Create DataFrame
-                param_df = pd.DataFrame(param_values)
-                
-                if len(param_df) > 0 and len(param_df.columns) > 2:  # Need more than just run_number and best_fitness
-                    # Calculate correlation matrix
-                    corr_matrix = param_df.corr()
-                    
-                    # Create heatmap
-                    sns.heatmap(corr_matrix, annot=True, cmap='coolwarm', fmt='.2f', 
-                               linewidths=0.5, ax=ax_heatmap, vmin=-1, vmax=1)
-                    ax_heatmap.set_title("Correlation Between Parameters and Fitness", fontsize=14)
-                    
-                    # Create canvas and add to layout
-                    canvas_heatmap = FigureCanvasQTAgg(fig_heatmap)
-                    self.heatmap_plot_widget.layout().addWidget(canvas_heatmap)
-                    
-                    # Add toolbar for interactive features
-                    toolbar_heatmap = NavigationToolbar(canvas_heatmap, self.heatmap_plot_widget)
-                    self.heatmap_plot_widget.layout().addWidget(toolbar_heatmap)
-
-                    # Add "Open in New Window" button
-                    open_new_window_button = QPushButton("Open in New Window")
-                    open_new_window_button.setObjectName("secondary-button")
-                    open_new_window_button.clicked.connect(lambda: self._open_plot_window(fig_heatmap, "Heatmap Plot"))
-                    self.heatmap_plot_widget.layout().addWidget(open_new_window_button)
-                else:
-                    # Create a label for insufficient data
-                    no_data_label = QLabel("Insufficient data for correlation analysis")
-                    self.heatmap_plot_widget.layout().addWidget(no_data_label)
-            else:
-                # Create a label if no parameter data
-                no_data_label = QLabel("No parameter data available for correlation analysis")
-                self.heatmap_plot_widget.layout().addWidget(no_data_label)
-        except Exception as e:
-            print(f"Error creating heatmap: {str(e)}")
-            error_label = QLabel(f"Error creating heatmap: {str(e)}")
-            self.heatmap_plot_widget.layout().addWidget(error_label)
-            
-        # 5. Create Q-Q plot for normality assessment
-        try:
-            # Clear existing plot layout
-            if self.qq_plot_widget.layout():
-                for i in reversed(range(self.qq_plot_widget.layout().count())): 
-                    self.qq_plot_widget.layout().itemAt(i).widget().setParent(None)
-            else:
-                self.qq_plot_widget.setLayout(QVBoxLayout())
-                
-            # Create figure for Q-Q plot
-            fig_qq = Figure(figsize=(10, 6), tight_layout=True)
-            ax_qq = fig_qq.add_subplot(111)
-            
-            # Create Q-Q plot
-            from scipy import stats
-            stats.probplot(df["best_fitness"], dist="norm", plot=ax_qq)
-            ax_qq.set_title("Q-Q Plot for Normality Assessment", fontsize=14)
-            ax_qq.set_xlabel("Theoretical Quantiles", fontsize=12)
-            ax_qq.set_ylabel("Sample Quantiles", fontsize=12)
-            ax_qq.grid(True, linestyle="--", alpha=0.7)
-            
-            # Perform normality tests
-            shapiro_test = stats.shapiro(df["best_fitness"])
-            ks_test = stats.kstest(df["best_fitness"], 'norm', 
-                                 args=(df["best_fitness"].mean(), df["best_fitness"].std()))
-            
-            # Add test results as text
-            test_text = (
-                f"Shapiro-Wilk Test:\n"
-                f"W = {shapiro_test[0]:.4f}\n"
-                f"p-value = {shapiro_test[1]:.4f}\n\n"
-                f"Kolmogorov-Smirnov Test:\n"
-                f"D = {ks_test[0]:.4f}\n"
-                f"p-value = {ks_test[1]:.4f}"
-            )
-            
-            # Create a text box for the test results
-            props = dict(boxstyle='round', facecolor='wheat', alpha=0.5)
-            ax_qq.text(0.05, 0.95, test_text, transform=ax_qq.transAxes, 
-                      fontsize=10, verticalalignment='top', bbox=props)
-            
-            # Create canvas and add to layout
-            canvas_qq = FigureCanvasQTAgg(fig_qq)
-            self.qq_plot_widget.layout().addWidget(canvas_qq)
-            
-            # Add toolbar for interactive features
-            toolbar_qq = NavigationToolbar(canvas_qq, self.qq_plot_widget)
-            self.qq_plot_widget.layout().addWidget(toolbar_qq)
-
-            # Add "Open in New Window" button
-            open_new_window_button = QPushButton("Open in New Window")
-            open_new_window_button.setObjectName("secondary-button")
-            open_new_window_button.clicked.connect(lambda: self._open_plot_window(fig_qq, "Q-Q Plot"))
-            self.qq_plot_widget.layout().addWidget(open_new_window_button)
-
-        except Exception as e:
-            print(f"Error creating Q-Q plot: {str(e)}")
-        
-        # 6. Update statistics table
-        try:
-            # Calculate statistics for fitness and available parameters
-            stats_data = []
-            
-            # Add fitness statistics
-            fitness_stats = {
-                "Metric": "Best Fitness",
-                "Min": df["best_fitness"].min(),
-                "Max": df["best_fitness"].max(),
-                "Mean": df["best_fitness"].mean(),
-                "Std": df["best_fitness"].std()
-            }
-            stats_data.append(fitness_stats)
-            
-            # Add statistics for other metrics in results
-            for col in df.columns:
-                if col not in ["run_number", "best_fitness", "best_solution", "parameter_names"] and df[col].dtype in [np.float64, np.int64]:
-                    metric_stats = {
-                        "Metric": col,
-                        "Min": df[col].min(),
-                        "Max": df[col].max(),
-                        "Mean": df[col].mean(),
-                        "Std": df[col].std()
-                    }
-                    stats_data.append(metric_stats)
-            
-            # Update table with statistics
-            self.benchmark_stats_table.setRowCount(len(stats_data))
-            for row, stat in enumerate(stats_data):
-                self.benchmark_stats_table.setItem(row, 0, QTableWidgetItem(str(stat["Metric"])))
-                self.benchmark_stats_table.setItem(row, 1, QTableWidgetItem(f"{stat['Min']:.6f}"))
-                self.benchmark_stats_table.setItem(row, 2, QTableWidgetItem(f"{stat['Max']:.6f}"))
-                self.benchmark_stats_table.setItem(row, 3, QTableWidgetItem(f"{stat['Mean']:.6f}"))
-                self.benchmark_stats_table.setItem(row, 4, QTableWidgetItem(f"{stat['Std']:.6f}"))
-                
-            # 7. Update runs table with fitness, rank and best/worst/mean indicators
-            self.benchmark_runs_table.setRowCount(len(df))
-            
-            # Sort runs by fitness (assuming lower is better)
-            sorted_df = df.sort_values('best_fitness')
-            
-            # Get index of run with fitness value closest to mean
-            mean_fitness = df['best_fitness'].mean()
-            mean_index = (df['best_fitness'] - mean_fitness).abs().idxmin()
-            
-            # Create a button class for the details button
-            class DetailButton(QPushButton):
-                def __init__(self, run_number):
-                    super().__init__("View Details")
-                    self.run_number = run_number
-            
-            # Populate the table
-            for i, (_, row) in enumerate(sorted_df.iterrows()):
-                run_number = int(row['run_number'])
-                fitness = row['best_fitness']
-                
-                # Create items for the table
-                run_item = QTableWidgetItem(str(run_number))
-                fitness_item = QTableWidgetItem(f"{fitness:.6f}")
-                rank_item = QTableWidgetItem(f"{i+1}/{len(df)}")
-                
-                # Set alignment
-                run_item.setTextAlignment(Qt.AlignCenter)
-                fitness_item.setTextAlignment(Qt.AlignCenter)
-                rank_item.setTextAlignment(Qt.AlignCenter)
-                
-                # Color coding
-                if i == 0:  # Best run (lowest fitness)
-                    run_item.setBackground(QColor(200, 255, 200))  # Light green
-                    fitness_item.setBackground(QColor(200, 255, 200))
-                    rank_item.setBackground(QColor(200, 255, 200))
-                    run_item.setToolTip("Best Run (Lowest Fitness)")
-                elif i == len(df) - 1:  # Worst run (highest fitness)
-                    run_item.setBackground(QColor(255, 200, 200))  # Light red
-                    fitness_item.setBackground(QColor(255, 200, 200))
-                    rank_item.setBackground(QColor(255, 200, 200))
-                    run_item.setToolTip("Worst Run (Highest Fitness)")
-                elif row.name == mean_index:  # Mean run (closest to mean fitness)
-                    run_item.setBackground(QColor(255, 255, 200))  # Light yellow
-                    fitness_item.setBackground(QColor(255, 255, 200))
-                    rank_item.setBackground(QColor(255, 255, 200))
-                    run_item.setToolTip("Mean Run (Closest to Average Fitness)")
-                
-                # Add items to the table
-                self.benchmark_runs_table.setItem(i, 0, run_item)
-                self.benchmark_runs_table.setItem(i, 1, fitness_item)
-                self.benchmark_runs_table.setItem(i, 2, rank_item)
-                
-                # Add a details button
-                detail_btn = DetailButton(run_number)
-                detail_btn.clicked.connect(lambda _, btn=detail_btn: self.show_run_details(
-                    self.benchmark_runs_table.item(
-                        [i for i in range(self.benchmark_runs_table.rowCount()) 
-                         if int(self.benchmark_runs_table.item(i, 0).text()) == btn.run_number][0], 0)))
-                self.benchmark_runs_table.setCellWidget(i, 3, detail_btn)
-        except Exception as e:
-            print(f"Error updating statistics tables: {str(e)}")
-        
-        # Connect export button if not already connected
-        try:
-            self.export_benchmark_button.clicked.disconnect()
-        except:
-            pass
-        self.export_benchmark_button.clicked.connect(self.export_ga_benchmark_data)
-        
-    def export_ga_benchmark_data(self):
-        """Export GA benchmark data to a JSON file with all visualization data"""
-        try:
-            import pandas as pd
-            import json
-            import numpy as np
-            from datetime import datetime
-            
-            # Create enhanced benchmark data with all necessary visualization metrics
-            enhanced_data = []
-            for run in self.ga_benchmark_data:
-                enhanced_run = run.copy()
-                
-                # Ensure benchmark_metrics exists and is a dictionary
-                if 'benchmark_metrics' not in enhanced_run or not isinstance(enhanced_run['benchmark_metrics'], dict):
-                    enhanced_run['benchmark_metrics'] = {}
-                
-                # Create synthetic data for missing metrics to ensure visualizations work
-                metrics = enhanced_run['benchmark_metrics']
-                
-                # Add essential metrics if missing
-                if not metrics.get('fitness_history'):
-                    # Create synthetic fitness history
-                    generations = 50  # Default number of generations
-                    if 'best_fitness_per_gen' in metrics and metrics['best_fitness_per_gen']:
-                        generations = len(metrics['best_fitness_per_gen'])
-                    else:
-                        # Create best fitness per generation
-                        best_fitness = enhanced_run.get('best_fitness', 1.0)
-                        metrics['best_fitness_per_gen'] = list(np.linspace(best_fitness * 2, best_fitness, generations))
-                    
-                    # Create fitness history - population fitness values for each generation
-                    pop_size = 100
-                    fitness_history = []
-                    for gen in range(generations):
-                        gen_fitness = []
-                        best_in_gen = metrics['best_fitness_per_gen'][gen]
-                        for i in range(pop_size):
-                            # Add some random variation
-                            gen_fitness.append(best_in_gen * (1 + np.random.rand() * 0.5))
-                        fitness_history.append(gen_fitness)
-                    metrics['fitness_history'] = fitness_history
-                
-                # Add mean fitness history if missing
-                if not metrics.get('mean_fitness_history') and metrics.get('fitness_history'):
-                    metrics['mean_fitness_history'] = [np.mean(gen) for gen in metrics['fitness_history']]
-                
-                # Add std fitness history if missing
-                if not metrics.get('std_fitness_history') and metrics.get('fitness_history'):
-                    metrics['std_fitness_history'] = [np.std(gen) for gen in metrics['fitness_history']]
-                
-                # Add parameter convergence data if missing
-                if (not metrics.get('best_individual_per_gen') and 
-                    metrics.get('best_fitness_per_gen') and 
-                    'best_solution' in enhanced_run and 
-                    'parameter_names' in enhanced_run):
-                    
-                    generations = len(metrics['best_fitness_per_gen'])
-                    final_solution = enhanced_run['best_solution']
-                    
-                    # Create parameter convergence data - parameters evolving towards final solution
-                    best_individual_per_gen = []
-                    for gen in range(generations):
-                        # Start with random values and gradually converge to final solution
-                        progress = gen / (generations - 1) if generations > 1 else 1
-                        gen_solution = []
-                        for param in final_solution:
-                            # Random initial value that converges to final
-                            initial = param * 2 if param != 0 else 0.5
-                            gen_solution.append(initial * (1 - progress) + param * progress)
-                        best_individual_per_gen.append(gen_solution)
-                    
-                    metrics['best_individual_per_gen'] = best_individual_per_gen
-                
-                # Add adaptive rates data if missing
-                if not metrics.get('adaptive_rates_history') and metrics.get('best_fitness_per_gen'):
-                    generations = len(metrics['best_fitness_per_gen'])
-                    
-                    # Create adaptive rates history
-                    adaptive_rates_history = []
-                    cxpb = 0.7  # Starting crossover probability
-                    mutpb = 0.2  # Starting mutation probability
-                    
-                    for gen in range(0, generations, max(1, generations // 10)):
-                        # Every few generations, adapt rates
-                        old_cxpb = cxpb
-                        old_mutpb = mutpb
-                        
-                        # Simple adaptation strategy
-                        if gen % 3 == 0:
-                            cxpb = min(0.9, cxpb + 0.05)
-                            mutpb = max(0.1, mutpb - 0.02)
-                            adaptation_type = "Exploration"
-                        else:
-                            cxpb = max(0.5, cxpb - 0.03)
-                            mutpb = min(0.5, mutpb + 0.03)
-                            adaptation_type = "Exploitation"
-                        
-                        adaptive_rates_history.append({
-                            'generation': gen,
-                            'old_cxpb': old_cxpb,
-                            'new_cxpb': cxpb,
-                            'old_mutpb': old_mutpb,
-                            'new_mutpb': mutpb,
-                            'adaptation_type': adaptation_type
-                        })
-                    
-                    metrics['adaptive_rates_history'] = adaptive_rates_history
-                
-                # Add computational metrics if missing
-                if not metrics.get('cpu_usage'):
-                    metrics['cpu_usage'] = list(10 + 70 * np.random.rand(100))
-                
-                if not metrics.get('memory_usage'):
-                    metrics['memory_usage'] = list(100 + 500 * np.random.rand(100))
-                
-                if not metrics.get('evaluation_times'):
-                    metrics['evaluation_times'] = list(0.05 + 0.02 * np.random.rand(50))
-                
-                if not metrics.get('crossover_times'):
-                    metrics['crossover_times'] = list(0.02 + 0.01 * np.random.rand(50))
-                
-                if not metrics.get('mutation_times'):
-                    metrics['mutation_times'] = list(0.01 + 0.005 * np.random.rand(50))
-                
-                if not metrics.get('selection_times'):
-                    metrics['selection_times'] = list(0.03 + 0.01 * np.random.rand(50))
-                
-                enhanced_data.append(enhanced_run)
-            
-            # Create a custom JSON encoder to handle NumPy types
-            class NumpyEncoder(json.JSONEncoder):
-                def default(self, obj):
-                    if isinstance(obj, np.ndarray):
-                        return obj.tolist()
-                    if isinstance(obj, np.integer):
-                        return int(obj)
-                    if isinstance(obj, np.floating):
-                        return float(obj)
-                    return json.JSONEncoder.default(self, obj)
-            
-            # Ask user for save location
-            file_path, _ = QFileDialog.getSaveFileName(
-                self, 
-                "Export GA Benchmark Data", 
-                f"ga_benchmark_data_{QDateTime.currentDateTime().toString('yyyyMMdd_hhmmss')}.json", 
-                "JSON Files (*.json);;All Files (*)"
-            )
-            
-            if not file_path:
-                return  # User cancelled
-                
-            # Add .json extension if not provided
-            if not file_path.lower().endswith('.json'):
-                file_path += '.json'
-            
-            # Add timestamp to data
-            export_data = {
-                'ga_benchmark_data': enhanced_data,
-                'export_timestamp': datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-            }
-            
-            # Save to file
-            with open(file_path, 'w') as f:
-                json.dump(export_data, f, indent=2, cls=NumpyEncoder)
-            
-            self.status_bar.showMessage(f"Enhanced benchmark data exported to {file_path}")
-            
-        except Exception as e:
-            QMessageBox.critical(self, "Export Error", f"Error exporting benchmark data: {str(e)}")
-            import traceback
-            print(f"Export error details: {traceback.format_exc()}")
-            
-    def import_ga_benchmark_data(self):
-        """Import GA benchmark data from a CSV file"""
-        try:
-            import pandas as pd
-            from PyQt5.QtWidgets import QFileDialog
-            
-            # Ask user for file location
-            file_path, _ = QFileDialog.getOpenFileName(
-                self, 
-                "Import GA Benchmark Data", 
-                "", 
-                "CSV Files (*.csv);;All Files (*)"
-            )
-            
-            if not file_path:
-                return  # User cancelled
-                
-            # Load from file
-            df = pd.read_csv(file_path)
-            
-            # Convert string representations back to lists for best_solution and parameter_names
-            if 'best_solution' in df.columns:
-                df['best_solution'] = df['best_solution'].apply(
-                    lambda x: [float(val) for val in x.split(';')] if isinstance(x, str) else x
-                )
-                
-            if 'parameter_names' in df.columns:
-                df['parameter_names'] = df['parameter_names'].apply(
-                    lambda x: x.split(';') if isinstance(x, str) else x
-                )
-            
-            # Convert DataFrame to list of dictionaries
-            self.ga_benchmark_data = df.to_dict('records')
-            
-            # Enable the export button
-            self.export_benchmark_button.setEnabled(True)
-            
-            # Update visualizations
-            self.visualize_ga_benchmark_results()
-            
-            self.status_bar.showMessage(f"Benchmark data imported from {file_path}")
-            
-        except Exception as e:
-            QMessageBox.critical(self, "Import Error", f"Error importing benchmark data: {str(e)}")
-            import traceback
-            print(f"Import error details: {traceback.format_exc()}")
-            
-    def show_run_details(self, item):
-        """Show detailed information about the selected benchmark run"""
-        if not hasattr(self, 'ga_benchmark_data') or not self.ga_benchmark_data:
-            return
-            
-        # Get row index of the clicked item
-        row = item.row()
-        
-        # Get run info from table
-        run_number_item = self.benchmark_runs_table.item(row, 0)
-        if not run_number_item:
-            return
-            
-        run_number_text = run_number_item.text()
-        try:
-            run_number = int(run_number_text)
-        except ValueError:
-            return
-            
-        # Find the run data
-        run_data = None
-        for run in self.ga_benchmark_data:
-            if run.get('run_number') == run_number:
-                run_data = run
-                break
-                
-        if not run_data:
-            self.run_details_text.setText("Run data not found.")
-            return
-            
-        # Build detailed information
-        details = []
-        details.append(f"<h3>Run #{run_number} Details</h3>")
-        details.append(f"<p><b>Best Fitness:</b> {run_data.get('best_fitness', 'N/A'):.6f}</p>")
-        
-        # Add any other metrics that might be available
-        for key, value in run_data.items():
-            if key not in ['run_number', 'best_fitness', 'best_solution', 'parameter_names'] and isinstance(value, (int, float)):
-                details.append(f"<p><b>{key}:</b> {value:.6f}</p>")
-                
-        # Add optimized DVA parameters
-        if 'best_solution' in run_data and 'parameter_names' in run_data:
-            details.append("<h4>Optimized DVA Parameters:</h4>")
-            details.append("<table border='1' cellspacing='0' cellpadding='5' style='border-collapse: collapse;'>")
-            details.append("<tr><th>Parameter</th><th>Value</th></tr>")
-            
-            solution = run_data['best_solution']
-            param_names = run_data['parameter_names']
-            
-            for i, (param, value) in enumerate(zip(param_names, solution)):
-                details.append(f"<tr><td>{param}</td><td>{value:.6f}</td></tr>")
-                
-            details.append("</table>")
-            
-        # Set the detailed text
-        self.run_details_text.setHtml("".join(details))
-        
-        try:
-            import pandas as pd
-            from PyQt5.QtWidgets import QVBoxLayout, QLabel
-            from computational_metrics_new import (
-                visualize_ga_operations, create_ga_visualizations, ensure_all_visualizations_visible
-            )
-            
-            # Create a DataFrame with just this run's data
-            run_df = pd.DataFrame([run_data])
-            
-            # We do NOT update the global visualization plots (CPU, memory, I/O)
-            # These should only show aggregate data for all runs
-            
-            # Handle GA operations widget - make sure all plots are properly displayed
-            if hasattr(self, 'ga_ops_plot_widget'):
-                # Clear the GA operations widget before visualizing
-                if self.ga_ops_plot_widget.layout():
-                    for i in reversed(range(self.ga_ops_plot_widget.layout().count())): 
-                        self.ga_ops_plot_widget.layout().itemAt(i).widget().setParent(None)
-                else:
-                    self.ga_ops_plot_widget.setLayout(QVBoxLayout())
-                    
-                # Print available data for debugging
-                print(f"Run data keys: {list(run_data.keys())}")
-                if 'benchmark_metrics' in run_data:
-                    print(f"Benchmark metrics type: {type(run_data['benchmark_metrics'])}")
-                    if isinstance(run_data['benchmark_metrics'], dict):
-                        print(f"Benchmark metrics keys: {list(run_data['benchmark_metrics'].keys())}")
-                
-                # Create tabs for different visualization types within GA operations
-                ga_ops_tabs = QTabWidget()
-                self.ga_ops_plot_widget.layout().addWidget(ga_ops_tabs)
-                
-                # Create tabs only for fitness evolution and adaptive rates
-                fitness_tab = QWidget()
-                fitness_tab.setLayout(QVBoxLayout())
-                rates_tab = QWidget()
-                rates_tab.setLayout(QVBoxLayout())
-                
-                # Add the tabs - only fitness evolution and adaptive rates
-                ga_ops_tabs.addTab(fitness_tab, "Fitness Evolution")
-                ga_ops_tabs.addTab(rates_tab, "Adaptive Rates")
-                
-                # Try to create each visualization in its own tab
-                try:
-                    # Create fitness evolution plot
-                    self.create_fitness_evolution_plot(fitness_tab, run_data)
-                    
-                    # Create adaptive rates plot
-                    self.create_adaptive_rates_plot(rates_tab, run_data)
-                except Exception as viz_error:
-                    print(f"Error in visualization tabs: {str(viz_error)}")
-                    try:
-                        # Fallback to basic visualization
-                        visualize_ga_operations(self.ga_ops_plot_widget, run_df)
-                    except Exception as basic_viz_error:
-                        print(f"Error in basic visualization: {str(basic_viz_error)}")
-                        # Add error message to widget
-                        if self.ga_ops_plot_widget.layout():
-                            self.ga_ops_plot_widget.layout().addWidget(QLabel(f"Error visualizing GA operations: {str(viz_error)}"))
-                
-                # Make sure all visualizations are visible
-                ensure_all_visualizations_visible(self.ga_ops_plot_widget)
-            
-            # Make sure all tabs in the main tab widget are preserved and properly displayed
-            if hasattr(self, 'benchmark_viz_tabs'):
-                # First, switch to the Statistics tab to make the details visible
-                stats_tab_index = self.benchmark_viz_tabs.indexOf(self.benchmark_viz_tabs.findChild(QWidget, "stats_tab"))
-                if stats_tab_index == -1:  # If not found by name, try finding by index
-                    stats_tab_index = 5  # Statistics tab is typically the 6th tab (index 5)
-                
-                # Switch to the stats tab
-                self.benchmark_viz_tabs.setCurrentIndex(stats_tab_index)
-                
-                # Make sure all tabs and their contents are visible
-                for i in range(self.benchmark_viz_tabs.count()):
-                    tab = self.benchmark_viz_tabs.widget(i)
-                    if tab:
-                        tab.setVisible(True)
-                        # If the tab has a layout, make all its children visible
-                        if tab.layout():
-                            for j in range(tab.layout().count()):
-                                child = tab.layout().itemAt(j).widget()
-                                if child:
-                                    child.setVisible(True)
-                
-                # We don't update the general visualization tabs, they should only show aggregate data
-        except Exception as e:
-            import traceback
-            print(f"Error visualizing run metrics: {str(e)}\n{traceback.format_exc()}")
-            
-    def update_all_visualizations(self, run_data):
-        """
-        Update all visualization tabs with the given run data.
-        This ensures that all plots are properly displayed when viewing run details.
-        
-        NOTE: This method is no longer used as per user requirement. The visualization plots
-        (violin, distribution, scatter, parameter correlation, QQ, CPU, memory, IO) 
-        should only show aggregate data for all runs, not individual run data.
-        
-        Args:
-            run_data: Dictionary containing the run data to visualize
-        """
-        try:
-            import pandas as pd
-            import numpy as np
-            import matplotlib.pyplot as plt
-            from matplotlib.figure import Figure
-            from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-            from computational_metrics_new import ensure_all_visualizations_visible
-            
-            # Create a DataFrame with just this run's data
-            run_df = pd.DataFrame([run_data])
-            
-            # Update Violin Plot if available
-            if hasattr(self, 'violin_plot_widget') and self.violin_plot_widget:
-                self.setup_widget_layout(self.violin_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # Create a violin plot for a single run (not very useful, but we can show something)
-                    if 'best_fitness' in run_data:
-                        ax.set_title(f"Fitness Value for Run #{run_data.get('run_number', 1)}")
-                        ax.set_ylabel("Fitness Value")
-                        ax.set_xticks([1])
-                        ax.set_xticklabels([f"Run #{run_data.get('run_number', 1)}"])
-                        ax.bar([1], [run_data['best_fitness']], width=0.6, alpha=0.7, color='blue')
-                        ax.text(1, run_data['best_fitness'], f"{run_data['best_fitness']:.6f}", 
-                                ha='center', va='bottom', fontsize=10)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.violin_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.violin_plot_widget)
-                except Exception as e:
-                    print(f"Error updating violin plot: {str(e)}")
-            
-            # Update Distribution Plot if available
-            if hasattr(self, 'dist_plot_widget') and self.dist_plot_widget:
-                self.setup_widget_layout(self.dist_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # For distribution of a single run, show parameter values
-                    if 'best_solution' in run_data and 'parameter_names' in run_data:
-                        solution = run_data['best_solution']
-                        param_names = run_data['parameter_names']
-                        
-                        # Only show non-zero parameters for clarity
-                        non_zero_params = [(name, val) for name, val in zip(param_names, solution) if abs(val) > 1e-6]
-                        
-                        if non_zero_params:
-                            names, values = zip(*non_zero_params)
-                            y_pos = range(len(names))
-                            
-                            # Create horizontal bar chart of parameter values
-                            ax.barh(y_pos, values, align='center', alpha=0.7, color='green')
-                            ax.set_yticks(y_pos)
-                            ax.set_yticklabels(names)
-                            ax.invert_yaxis()  # Labels read top-to-bottom
-                            ax.set_xlabel('Parameter Value')
-                            ax.set_title('Non-Zero Parameter Values for Selected Run')
-                            
-                            # Add value labels
-                            for i, v in enumerate(values):
-                                ax.text(v + 0.01, i, f"{v:.4f}", va='center')
-                        else:
-                            ax.text(0.5, 0.5, "No non-zero parameters found", 
-                                   ha='center', va='center', transform=ax.transAxes)
-                    else:
-                        ax.text(0.5, 0.5, "No parameter data available", 
-                               ha='center', va='center', transform=ax.transAxes)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.dist_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.dist_plot_widget)
-                except Exception as e:
-                    print(f"Error updating distribution plot: {str(e)}")
-            
-            # Update Scatter Plot if available
-            if hasattr(self, 'scatter_plot_widget') and self.scatter_plot_widget:
-                self.setup_widget_layout(self.scatter_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # For a scatter plot of a single run, show fitness history if available
-                    if 'benchmark_metrics' in run_data and isinstance(run_data['benchmark_metrics'], dict):
-                        metrics = run_data['benchmark_metrics']
-                        if 'fitness_history' in metrics and metrics['fitness_history']:
-                            # Get fitness history for each generation
-                            generations = range(1, len(metrics['fitness_history']) + 1)
-                            best_fitness_per_gen = [min(gen_fitness) if gen_fitness else float('nan') 
-                                                   for gen_fitness in metrics['fitness_history']]
-                            
-                            # Plot fitness evolution
-                            ax.plot(generations, best_fitness_per_gen, 'b-', marker='o', markersize=4, linewidth=2)
-                            ax.set_xlabel('Generation')
-                            ax.set_ylabel('Best Fitness')
-                            ax.set_title(f'Fitness Evolution for Run #{run_data.get("run_number", 1)}')
-                            ax.grid(True, linestyle='--', alpha=0.7)
-                        else:
-                            ax.text(0.5, 0.5, "No fitness history available", 
-                                   ha='center', va='center', transform=ax.transAxes)
-                    else:
-                        ax.text(0.5, 0.5, "No benchmark metrics available", 
-                               ha='center', va='center', transform=ax.transAxes)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.scatter_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.scatter_plot_widget)
-                except Exception as e:
-                    print(f"Error updating scatter plot: {str(e)}")
-            
-            # Update Heatmap Plot if available
-            if hasattr(self, 'heatmap_plot_widget') and self.heatmap_plot_widget:
-                self.setup_widget_layout(self.heatmap_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # For a heatmap of a single run, show parameter correlations
-                    if 'best_solution' in run_data and 'parameter_names' in run_data:
-                        solution = run_data['best_solution']
-                        param_names = run_data['parameter_names']
-                        
-                        # Create a mock correlation matrix (not real correlations for a single run)
-                        # Just show which parameters are active
-                        active_params = [i for i, val in enumerate(solution) if abs(val) > 1e-6]
-                        active_names = [param_names[i] for i in active_params]
-                        
-                        if active_params:
-                            # Create a matrix showing active parameters
-                            n = len(active_params)
-                            matrix = np.ones((n, n))
-                            
-                            # Create heatmap
-                            im = ax.imshow(matrix, cmap='viridis')
-                            
-                            # Set ticks and labels
-                            ax.set_xticks(range(n))
-                            ax.set_yticks(range(n))
-                            ax.set_xticklabels(active_names, rotation=90)
-                            ax.set_yticklabels(active_names)
-                            
-                            # Add text showing parameter values
-                            for i in range(n):
-                                for j in range(n):
-                                    val = solution[active_params[i]]
-                                    text = f"{val:.3f}" if i == j else ""
-                                    ax.text(j, i, text, ha="center", va="center", 
-                                           color="white" if matrix[i, j] > 0.5 else "black")
-                            
-                            ax.set_title("Active Parameters in Solution")
-                            fig.colorbar(im, ax=ax, label="Parameter Active")
-                        else:
-                            ax.text(0.5, 0.5, "No active parameters found", 
-                                   ha='center', va='center', transform=ax.transAxes)
-                    else:
-                        ax.text(0.5, 0.5, "No parameter data available", 
-                               ha='center', va='center', transform=ax.transAxes)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.heatmap_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.heatmap_plot_widget)
-                except Exception as e:
-                    print(f"Error updating heatmap plot: {str(e)}")
-            
-            # Update Q-Q Plot if available
-            if hasattr(self, 'qq_plot_widget') and self.qq_plot_widget:
-                self.setup_widget_layout(self.qq_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # For a Q-Q plot of a single run, we can't do much, so show a message
-                    ax.text(0.5, 0.5, "Q-Q plot requires multiple runs for comparison", 
-                           ha='center', va='center', transform=ax.transAxes, fontsize=12)
-                    ax.set_title("Q-Q Plot")
-                    ax.set_xlabel("Theoretical Quantiles")
-                    ax.set_ylabel("Sample Quantiles")
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.qq_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.qq_plot_widget)
-                except Exception as e:
-                    print(f"Error updating Q-Q plot: {str(e)}")
-            
-        except Exception as e:
-            import traceback
-            print(f"Error updating all visualizations: {str(e)}\n{traceback.format_exc()}")
-    
-    def setup_widget_layout(self, widget):
-        """
-        Clear existing layout or create a new one for a widget
-        
-        Args:
-            widget: QWidget to set up layout for
-        """
-        if widget.layout():
-            # Clear existing layout
-            for i in reversed(range(widget.layout().count())): 
-                widget.layout().itemAt(i).widget().setParent(None)
-        else:
-            # Create new layout
-            widget.setLayout(QVBoxLayout())
-            
-    def create_fitness_evolution_plot(self, tab_widget, run_data):
-        """
-        Create a fitness evolution plot in the specified tab widget
-        
-        Args:
-            tab_widget: Widget to place the plot in
-            run_data: Dictionary containing run data
-        """
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from matplotlib.figure import Figure
-        from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg, NavigationToolbar2QT
-        from computational_metrics_new import ensure_all_visualizations_visible
-        
-        # Create figure for fitness evolution with constrained size to prevent window expansion
-        fig = Figure(figsize=(7, 4), tight_layout=True)
-        ax = fig.add_subplot(111)
-        
-        # Get data
-        metrics = {}
-        if 'benchmark_metrics' in run_data and isinstance(run_data['benchmark_metrics'], dict):
-            metrics = run_data['benchmark_metrics']
-        
-        # Extract fitness history data
-        fitness_history = metrics.get('fitness_history', [])
-        mean_fitness_history = metrics.get('mean_fitness_history', [])
-        best_fitness_per_gen = metrics.get('best_fitness_per_gen', [])
-        
-        if best_fitness_per_gen:
-            # Plot data
-            generations = range(1, len(best_fitness_per_gen) + 1)
-            ax.plot(generations, best_fitness_per_gen, 'b-', linewidth=2, 
-                   label='Best Fitness')
-            
-            # Plot mean fitness if available
-            if mean_fitness_history and len(mean_fitness_history) == len(best_fitness_per_gen):
-                ax.plot(generations, mean_fitness_history, 'g-', linewidth=2, 
-                       alpha=0.7, label='Mean Fitness')
-            
-            # Add annotations
-            final_fitness = best_fitness_per_gen[-1]
-            percent_improvement = 0
-            if len(best_fitness_per_gen) > 1 and best_fitness_per_gen[0] != 0:
-                percent_improvement = ((best_fitness_per_gen[0] - final_fitness) / best_fitness_per_gen[0]) * 100
-            
-            # Add text box with summary
-            converge_text = f"Final fitness: {final_fitness:.6f}\nImprovement: {percent_improvement:.2f}%"
-            props = dict(boxstyle='round', facecolor='wheat', alpha=0.5)
-            ax.text(0.05, 0.95, converge_text, transform=ax.transAxes,
-                  fontsize=10, verticalalignment='top', bbox=props)
-                  
-            # Set labels and grid
-            ax.set_title("Fitness Evolution Over Generations", fontsize=14)
-            ax.set_xlabel("Generation", fontsize=12)
-            ax.set_ylabel("Fitness Value", fontsize=12)
-            ax.grid(True, linestyle="--", alpha=0.7)
-            ax.legend()
-        else:
-            ax.text(0.5, 0.5, "No fitness evolution data available", 
-                   ha='center', va='center', transform=ax.transAxes)
-        
-        # Add to widget
-        canvas = FigureCanvasQTAgg(fig)
-        tab_widget.layout().addWidget(canvas)
-        
-        # Add toolbar
-        toolbar = NavigationToolbar2QT(canvas, tab_widget)
-        tab_widget.layout().addWidget(toolbar)
-        
-        # Ensure visibility
-        ensure_all_visualizations_visible(tab_widget)
-        
-    def create_parameter_convergence_plot(self, tab_widget, run_data):
-        """
-        Create a parameter convergence plot in the specified tab widget
-        
-        Args:
-            tab_widget: Widget to place the plot in
-            run_data: Dictionary containing run data
-        """
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from matplotlib.figure import Figure
-        from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg, NavigationToolbar2QT
-        from computational_metrics_new import ensure_all_visualizations_visible
-        
-        # Create figure for parameter convergence with constrained size
-        fig = Figure(figsize=(7, 4), tight_layout=True)
-        ax = fig.add_subplot(111)
-        
-        # Get data
-        metrics = {}
-        if 'benchmark_metrics' in run_data and isinstance(run_data['benchmark_metrics'], dict):
-            metrics = run_data['benchmark_metrics']
-        
-        # Check for parameter data
-        best_individual_per_gen = metrics.get('best_individual_per_gen', [])
-        parameter_names = run_data.get('parameter_names', [])
-        
-        if best_individual_per_gen and parameter_names and len(best_individual_per_gen) > 0:
-            # Convert to numpy array for easier processing
-            param_array = np.array(best_individual_per_gen)
-            generations = range(1, len(best_individual_per_gen) + 1)
-            
-            # Find active parameters (non-zero values)
-            param_means = np.mean(param_array, axis=0)
-            active_params = np.where(param_means > 1e-6)[0]
-            
-            # If too many parameters, select most significant ones
-            if len(active_params) > 8:
-                param_ranges = np.max(param_array[:, active_params], axis=0) - np.min(param_array[:, active_params], axis=0)
-                significant_indices = np.argsort(param_ranges)[-8:]  # Take 8 most changing parameters
-                active_params = active_params[significant_indices]
-            
-            if len(active_params) > 0:
-                # Plot parameter convergence for active parameters
-                for i in active_params:
-                    if i < len(parameter_names):
-                        param_name = parameter_names[i]
-                        ax.plot(generations, param_array[:, i], label=param_name)
-                
-                # Set labels and grid
-                ax.set_title("Parameter Convergence Over Generations", fontsize=14)
-                ax.set_xlabel("Generation", fontsize=12)
-                ax.set_ylabel("Parameter Value", fontsize=12)
-                ax.grid(True, linestyle="--", alpha=0.7)
-                
-                # Add legend with smaller font to accommodate more parameters
-                ax.legend(fontsize=8, loc='upper center', bbox_to_anchor=(0.5, -0.15),
-                          fancybox=True, shadow=True, ncol=min(4, max(1, len(active_params))))
-                
-                fig.subplots_adjust(bottom=0.2)  # Make room for legend
-            else:
-                ax.text(0.5, 0.5, "No active parameters found", 
-                       ha='center', va='center', transform=ax.transAxes)
-        else:
-            ax.text(0.5, 0.5, "No parameter convergence data available", 
-                   ha='center', va='center', transform=ax.transAxes)
-        
-        # Add to widget
-        canvas = FigureCanvasQTAgg(fig)
-        tab_widget.layout().addWidget(canvas)
-        
-        # Add toolbar
-        toolbar = NavigationToolbar2QT(canvas, tab_widget)
-        tab_widget.layout().addWidget(toolbar)
-        
-        # Ensure visibility
-        ensure_all_visualizations_visible(tab_widget)
-        
-    def create_adaptive_rates_plot(self, tab_widget, run_data):
-        """
-        Create an adaptive rates plot in the specified tab widget
-        
-        Args:
-            tab_widget: Widget to place the plot in
-            run_data: Dictionary containing run data
-        """
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from matplotlib.figure import Figure
-        from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg, NavigationToolbar2QT
-        from computational_metrics_new import ensure_all_visualizations_visible
-        
-        # Create figure for adaptive rates with constrained size
-        fig = Figure(figsize=(7, 4), tight_layout=True)
-        ax = fig.add_subplot(111)
-        
-        # Get data
-        metrics = {}
-        if 'benchmark_metrics' in run_data and isinstance(run_data['benchmark_metrics'], dict):
-            metrics = run_data['benchmark_metrics']
-        
-        # Check for adaptive rates data
-        adaptive_rates_history = metrics.get('adaptive_rates_history', [])
-        
-        if adaptive_rates_history and len(adaptive_rates_history) > 0:
-            # Extract data
-            generations = [entry.get('generation', i) for i, entry in enumerate(adaptive_rates_history)]
-            old_cxpb = [entry.get('old_cxpb', 0) for entry in adaptive_rates_history]
-            new_cxpb = [entry.get('new_cxpb', 0) for entry in adaptive_rates_history]
-            old_mutpb = [entry.get('old_mutpb', 0) for entry in adaptive_rates_history]
-            new_mutpb = [entry.get('new_mutpb', 0) for entry in adaptive_rates_history]
-            
-            # Plot adaptive rates
-            ax.plot(generations, old_cxpb, 'b--', alpha=0.5, label='Old Crossover')
-            ax.plot(generations, new_cxpb, 'b-', linewidth=2, label='New Crossover')
-            ax.plot(generations, old_mutpb, 'r--', alpha=0.5, label='Old Mutation')
-            ax.plot(generations, new_mutpb, 'r-', linewidth=2, label='New Mutation')
-            
-            # Add annotations for adaptation type
-            for i, entry in enumerate(adaptive_rates_history):
-                adaptation_type = entry.get('adaptation_type', '')
-                if adaptation_type and i < len(generations):
-                    # Add a marker
-                    ax.plot(generations[i], new_cxpb[i], 'bo', markersize=6)
-                    ax.plot(generations[i], new_mutpb[i], 'ro', markersize=6)
-                    
-                    # Add annotation for every 3rd point to avoid clutter
-                    if i % 3 == 0:
-                        ax.annotate(adaptation_type.split('(')[0],
-                                   xy=(generations[i], new_cxpb[i]),
-                                   xytext=(10, 10),
-                                   textcoords='offset points',
-                                   arrowprops=dict(arrowstyle='->', connectionstyle='arc3,rad=.2'))
-            
-            # Set labels and grid
-            ax.set_title("Adaptive Rates During Optimization", fontsize=14)
-            ax.set_xlabel("Generation", fontsize=12)
-            ax.set_ylabel("Rate Value", fontsize=12)
-            ax.grid(True, linestyle="--", alpha=0.7)
-            ax.legend(loc='best')
-        else:
-            ax.text(0.5, 0.5, "No adaptive rates data available", 
-                   ha='center', va='center', transform=ax.transAxes)
-        
-        # Add to widget
-        canvas = FigureCanvasQTAgg(fig)
-        tab_widget.layout().addWidget(canvas)
-        
-        # Add toolbar
-        toolbar = NavigationToolbar2QT(canvas, tab_widget)
-        tab_widget.layout().addWidget(toolbar)
-        
-        # Ensure visibility
-        ensure_all_visualizations_visible(tab_widget)
-        
-    def create_computational_efficiency_plot(self, tab_widget, run_data):
-        """
-        Create a computational efficiency plot in the specified tab widget
-        
-        Args:
-            tab_widget: Widget to place the plot in
-            run_data: Dictionary containing run data
-        """
-        import numpy as np
-        import matplotlib.pyplot as plt
-        from matplotlib.figure import Figure
-        from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg, NavigationToolbar2QT
-        from computational_metrics_new import ensure_all_visualizations_visible
-        
-        # Create figure for computational efficiency with constrained size
-        fig = Figure(figsize=(7, 4), tight_layout=True)
-        ax = fig.add_subplot(111)
-        
-        # Get data
-        metrics = {}
-        if 'benchmark_metrics' in run_data and isinstance(run_data['benchmark_metrics'], dict):
-            metrics = run_data['benchmark_metrics']
-        
-        # Extract relevant metrics
-        cpu_usage = metrics.get('cpu_usage', [])
-        memory_usage = metrics.get('memory_usage', [])
-        evaluation_times = metrics.get('evaluation_times', [])
-        crossover_times = metrics.get('crossover_times', [])
-        mutation_times = metrics.get('mutation_times', [])
-        selection_times = metrics.get('selection_times', [])
-        
-        # Create a grid layout for multiple plots
-        from matplotlib.gridspec import GridSpec
-        gs = GridSpec(2, 2, figure=fig)
-        ax1 = fig.add_subplot(gs[0, 0])
-        ax2 = fig.add_subplot(gs[0, 1])
-        ax3 = fig.add_subplot(gs[1, 0])
-        ax4 = fig.add_subplot(gs[1, 1])
-        
-        # Plot 1: CPU Usage Over Time
-        if cpu_usage:
-            time_points = range(len(cpu_usage))
-            ax1.plot(time_points, cpu_usage, 'b-', linewidth=2)
-            ax1.set_title("CPU Usage During Optimization", fontsize=12)
-            ax1.set_xlabel("Time Point", fontsize=10)
-            ax1.set_ylabel("CPU Usage (%)", fontsize=10)
-            ax1.grid(True, linestyle="--", alpha=0.7)
-        else:
-            ax1.text(0.5, 0.5, "No CPU usage data available", 
-                   ha='center', va='center', transform=ax1.transAxes)
-        
-        # Plot 2: Scatter plot of CPU vs Fitness
-        if cpu_usage and metrics.get('best_fitness_per_gen', []):
-            best_fitness = metrics.get('best_fitness_per_gen', [])
-            # If different lengths, sample points
-            if len(cpu_usage) != len(best_fitness):
-                if len(cpu_usage) > len(best_fitness):
-                    # Sample CPU points
-                    points = np.linspace(0, len(cpu_usage)-1, len(best_fitness), dtype=int)
-                    sampled_cpu = [cpu_usage[i] for i in points]
-                    best_fitness_sample = best_fitness
-                else:
-                    # Sample fitness points
-                    points = np.linspace(0, len(best_fitness)-1, len(cpu_usage), dtype=int)
-                    best_fitness_sample = [best_fitness[i] for i in points]
-                    sampled_cpu = cpu_usage
-            else:
-                sampled_cpu = cpu_usage
-                best_fitness_sample = best_fitness
-            
-            # Create scatter plot
-            sc = ax2.scatter(sampled_cpu, best_fitness_sample, 
-                          c=range(len(sampled_cpu)), cmap='viridis',
-                          alpha=0.7, s=30)
-            fig.colorbar(sc, ax=ax2, label='Time Point')
-            ax2.set_title("CPU Usage vs. Fitness", fontsize=12)
-            ax2.set_xlabel("CPU Usage (%)", fontsize=10)
-            ax2.set_ylabel("Best Fitness", fontsize=10)
-            ax2.grid(True, linestyle="--", alpha=0.7)
-        else:
-            ax2.text(0.5, 0.5, "Insufficient data for CPU vs Fitness plot", 
-                   ha='center', va='center', transform=ax2.transAxes)
-        
-        # Plot 3: Memory Usage Over Time
-        if memory_usage:
-            time_points = range(len(memory_usage))
-            ax3.plot(time_points, memory_usage, 'g-', linewidth=2)
-            ax3.set_title("Memory Usage Over Time", fontsize=12)
-            ax3.set_xlabel("Time Point", fontsize=10)
-            ax3.set_ylabel("Memory Usage (MB)", fontsize=10)
-            ax3.grid(True, linestyle="--", alpha=0.7)
-        else:
-            ax3.text(0.5, 0.5, "No memory usage data available", 
-                   ha='center', va='center', transform=ax3.transAxes)
-        
-        # Plot 4: Operation Times
-        if any([evaluation_times, crossover_times, mutation_times, selection_times]):
-            # Compute average times per operation
-            op_names = []
-            op_times = []
-            
-            if evaluation_times:
-                op_names.append('Evaluation')
-                op_times.append(np.mean(evaluation_times))
-            if crossover_times:
-                op_names.append('Crossover')
-                op_times.append(np.mean(crossover_times))
-            if mutation_times:
-                op_names.append('Mutation')
-                op_times.append(np.mean(mutation_times))
-            if selection_times:
-                op_names.append('Selection')
-                op_times.append(np.mean(selection_times))
-            
-            # Create bar chart
-            if op_names and op_times:
-                ax4.bar(op_names, op_times, color='purple', alpha=0.7)
-                ax4.set_title("Average Operation Times", fontsize=12)
-                ax4.set_ylabel("Time (s)", fontsize=10)
-                ax4.grid(True, axis='y', linestyle="--", alpha=0.7)
-                
-                # Add values on top of bars
-                for i, v in enumerate(op_times):
-                    ax4.text(i, v + 0.001, f"{v:.3f}s", ha='center', fontsize=8)
-            else:
-                ax4.text(0.5, 0.5, "No operation time data available", 
-                       ha='center', va='center', transform=ax4.transAxes)
-        else:
-            ax4.text(0.5, 0.5, "No operation time data available", 
-                   ha='center', va='center', transform=ax4.transAxes)
-        
-        # Adjust layout
-        fig.tight_layout()
-        
-        # Add to widget
-        canvas = FigureCanvasQTAgg(fig)
-        tab_widget.layout().addWidget(canvas)
-        
-        # Add toolbar
-        toolbar = NavigationToolbar2QT(canvas, tab_widget)
-        tab_widget.layout().addWidget(toolbar)
-        
-        # Ensure visibility
-        ensure_all_visualizations_visible(tab_widget)
-    
-    def update_pso_visualizations(self, run_data):
-        """
-        Update all PSO visualization tabs with the given run data.
-        This ensures that all plots are properly displayed when viewing PSO run details.
-        
-        Args:
-            run_data: Dictionary containing the run data to visualize
-        """
-        try:
-            import pandas as pd
-            import numpy as np
-            import matplotlib.pyplot as plt
-            from matplotlib.figure import Figure
-            from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-            from computational_metrics_new import ensure_all_visualizations_visible
-            
-            # Create a DataFrame with just this run's data
-            run_df = pd.DataFrame([run_data])
-            
-            # Update PSO Violin Plot if available
-            if hasattr(self, 'pso_violin_plot_widget') and self.pso_violin_plot_widget:
-                self.setup_widget_layout(self.pso_violin_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # Create a violin plot for a single run (not very useful, but we can show something)
-                    if 'best_fitness' in run_data:
-                        ax.set_title(f"Fitness Value for PSO Run #{run_data.get('run_number', 1)}")
-                        ax.set_ylabel("Fitness Value")
-                        ax.set_xticks([1])
-                        ax.set_xticklabels([f"Run #{run_data.get('run_number', 1)}"])
-                        ax.bar([1], [run_data['best_fitness']], width=0.6, alpha=0.7, color='blue')
-                        ax.text(1, run_data['best_fitness'], f"{run_data['best_fitness']:.6f}", 
-                                ha='center', va='bottom', fontsize=10)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.pso_violin_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.pso_violin_plot_widget)
-                except Exception as e:
-                    print(f"Error updating PSO violin plot: {str(e)}")
-            
-            # Update PSO Distribution Plot if available
-            if hasattr(self, 'pso_dist_plot_widget') and self.pso_dist_plot_widget:
-                self.setup_widget_layout(self.pso_dist_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # For distribution of a single run, show parameter values
-                    if 'best_solution' in run_data and 'parameter_names' in run_data:
-                        solution = run_data['best_solution']
-                        param_names = run_data['parameter_names']
-                        
-                        # Only show non-zero parameters for clarity
-                        non_zero_params = [(name, val) for name, val in zip(param_names, solution) if abs(val) > 1e-6]
-                        
-                        if non_zero_params:
-                            names, values = zip(*non_zero_params)
-                            y_pos = range(len(names))
-                            
-                            # Create horizontal bar chart of parameter values
-                            ax.barh(y_pos, values, align='center', alpha=0.7, color='green')
-                            ax.set_yticks(y_pos)
-                            ax.set_yticklabels(names)
-                            ax.invert_yaxis()  # Labels read top-to-bottom
-                            ax.set_xlabel('Parameter Value')
-                            ax.set_title('Non-Zero Parameter Values for Selected PSO Run')
-                            
-                            # Add value labels
-                            for i, v in enumerate(values):
-                                ax.text(v + 0.01, i, f"{v:.4f}", va='center')
-                        else:
-                            ax.text(0.5, 0.5, "No non-zero parameters found", 
-                                   ha='center', va='center', transform=ax.transAxes)
-                    else:
-                        ax.text(0.5, 0.5, "No parameter data available", 
-                               ha='center', va='center', transform=ax.transAxes)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.pso_dist_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.pso_dist_plot_widget)
-                except Exception as e:
-                    print(f"Error updating PSO distribution plot: {str(e)}")
-            
-            # Update PSO Scatter Plot if available
-            if hasattr(self, 'pso_scatter_plot_widget') and self.pso_scatter_plot_widget:
-                self.setup_widget_layout(self.pso_scatter_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # For PSO scatter plot, show iterations vs fitness if available
-                    if 'optimization_metadata' in run_data and isinstance(run_data['optimization_metadata'], dict):
-                        metadata = run_data['optimization_metadata']
-                        if 'iterations' in metadata:
-                            iterations = metadata['iterations']
-                            # Create synthetic iteration data if needed
-                            iteration_points = range(1, iterations + 1)
-                            fitness_progress = np.linspace(run_data.get('best_fitness', 1.0) * 2, 
-                                                        run_data.get('best_fitness', 1.0), 
-                                                        iterations)
-                            
-                            # Plot fitness evolution
-                            ax.plot(iteration_points, fitness_progress, 'b-', marker='o', markersize=4, linewidth=2)
-                            ax.set_xlabel('Iteration')
-                            ax.set_ylabel('Best Fitness')
-                            ax.set_title(f'Fitness Evolution for PSO Run #{run_data.get("run_number", 1)}')
-                            ax.grid(True, linestyle='--', alpha=0.7)
-                        else:
-                            ax.text(0.5, 0.5, "No iteration data available", 
-                                   ha='center', va='center', transform=ax.transAxes)
-                    else:
-                        ax.text(0.5, 0.5, "No optimization metadata available", 
-                               ha='center', va='center', transform=ax.transAxes)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.pso_scatter_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.pso_scatter_plot_widget)
-                except Exception as e:
-                    print(f"Error updating PSO scatter plot: {str(e)}")
-            
-            # Update PSO Heatmap Plot if available
-            if hasattr(self, 'pso_heatmap_plot_widget') and self.pso_heatmap_plot_widget:
-                self.setup_widget_layout(self.pso_heatmap_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # Similar approach to GA heatmap
-                    if 'best_solution' in run_data and 'parameter_names' in run_data:
-                        solution = run_data['best_solution']
-                        param_names = run_data['parameter_names']
-                        
-                        # Create a mock correlation matrix
-                        active_params = [i for i, val in enumerate(solution) if abs(val) > 1e-6]
-                        active_names = [param_names[i] for i in active_params]
-                        
-                        if active_params:
-                            n = len(active_params)
-                            matrix = np.ones((n, n))
-                            
-                            # Create heatmap
-                            im = ax.imshow(matrix, cmap='viridis')
-                            
-                            # Set ticks and labels
-                            ax.set_xticks(range(n))
-                            ax.set_yticks(range(n))
-                            ax.set_xticklabels(active_names, rotation=90)
-                            ax.set_yticklabels(active_names)
-                            
-                            # Add text showing parameter values
-                            for i in range(n):
-                                for j in range(n):
-                                    val = solution[active_params[i]]
-                                    text = f"{val:.3f}" if i == j else ""
-                                    ax.text(j, i, text, ha="center", va="center", 
-                                           color="white" if matrix[i, j] > 0.5 else "black")
-                            
-                            ax.set_title("Active Parameters in PSO Solution")
-                            fig.colorbar(im, ax=ax, label="Parameter Active")
-                        else:
-                            ax.text(0.5, 0.5, "No active parameters found", 
-                                   ha='center', va='center', transform=ax.transAxes)
-                    else:
-                        ax.text(0.5, 0.5, "No parameter data available", 
-                               ha='center', va='center', transform=ax.transAxes)
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.pso_heatmap_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.pso_heatmap_plot_widget)
-                except Exception as e:
-                    print(f"Error updating PSO heatmap plot: {str(e)}")
-            
-            # Update PSO Q-Q Plot if available
-            if hasattr(self, 'pso_qq_plot_widget') and self.pso_qq_plot_widget:
-                self.setup_widget_layout(self.pso_qq_plot_widget)
-                try:
-                    fig = Figure(figsize=(7, 4), tight_layout=True)
-                    ax = fig.add_subplot(111)
-                    
-                    # Same message as for GA
-                    ax.text(0.5, 0.5, "Q-Q plot requires multiple runs for comparison", 
-                           ha='center', va='center', transform=ax.transAxes, fontsize=12)
-                    ax.set_title("Q-Q Plot")
-                    ax.set_xlabel("Theoretical Quantiles")
-                    ax.set_ylabel("Sample Quantiles")
-                        
-                    canvas = FigureCanvasQTAgg(fig)
-                    self.pso_qq_plot_widget.layout().addWidget(canvas)
-                    ensure_all_visualizations_visible(self.pso_qq_plot_widget)
-                except Exception as e:
-                    print(f"Error updating PSO Q-Q plot: {str(e)}")
-            
-        except Exception as e:
-            import traceback
-            print(f"Error updating all PSO visualizations: {str(e)}\n{traceback.format_exc()}")
-            
-    def export_parameters(self):
-        """Export main system and DVA parameters to a JSON file"""
-        try:
-            import json
-            from datetime import datetime
-            
-            # Get the parameters
-            main_params = self.get_main_system_params()
-            dva_params = self.get_dva_params()
-            
-            # Create a dictionary with all parameters
-            params_dict = {
-                "main_system": {
-                    "mu": main_params[0],
-                    "lambda": list(main_params[1:6]),
-                    "nu": list(main_params[6:11]),
-                    "a_low": main_params[11],
-                    "a_up": main_params[12],
-                    "f_1": main_params[13],
-                    "f_2": main_params[14],
-                    "omega_dc": main_params[15],
-                    "zeta_dc": main_params[16]
-                },
-                "dva": dva_params,
-                "timestamp": datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-            }
-            
-            # Ask user for save location
-            from PyQt5.QtWidgets import QFileDialog
-            file_path, _ = QFileDialog.getSaveFileName(
-                self, 
-                "Export Parameters", 
-                "", 
-                "JSON Files (*.json);;All Files (*)"
-            )
-            
-            if not file_path:
-                return  # User cancelled
-                
-            # Add .json extension if not provided
-            if not file_path.lower().endswith('.json'):
-                file_path += '.json'
-                
-            # Save to file
-            with open(file_path, 'w') as f:
-                json.dump(params_dict, f, indent=4)
-                
-            self.status_bar.showMessage(f"Parameters exported to {file_path}")
-            
-        except Exception as e:
-            QMessageBox.critical(self, "Export Error", f"Error exporting parameters: {str(e)}")
-    
-    def import_parameters(self):
-        """Import main system and DVA parameters from a JSON file"""
-        try:
-            import json
-            
-            # Ask user for file location
-            from PyQt5.QtWidgets import QFileDialog
-            file_path, _ = QFileDialog.getOpenFileName(
-                self, 
-                "Import Parameters", 
-                "", 
-                "JSON Files (*.json);;All Files (*)"
-            )
-            
-            if not file_path:
-                return  # User cancelled
-                
-            # Load from file
-            with open(file_path, 'r') as f:
-                params_dict = json.load(f)
-                
-            # Validate the data structure
-            if not isinstance(params_dict, dict) or "main_system" not in params_dict or "dva" not in params_dict:
-                raise ValueError("Invalid parameter file format")
-                
-            # Extract parameters
-            main_system = params_dict["main_system"]
-            dva = params_dict["dva"]
-            
-            # Format main system parameters into the expected tuple format
-            main_params = [
-                main_system.get("mu", 1.0),
-                *main_system.get("lambda", [1.0, 1.0, 0.5, 0.5, 0.5]),
-                *main_system.get("nu", [0.75, 0.75, 0.75, 0.75, 0.75]),
-                main_system.get("a_low", 0.05),
-                main_system.get("a_up", 0.05),
-                main_system.get("f_1", 100.0),
-                main_system.get("f_2", 100.0),
-                main_system.get("omega_dc", 5000.0),
-                main_system.get("zeta_dc", 0.01)
-            ]
-            
-            # Set main system parameters
-            self.set_main_system_params(main_params)
-            
-            # Set DVA parameters
-            self.set_dva_params(dva)
-            
-            self.status_bar.showMessage(f"Parameters imported from {file_path}")
-            
-        except Exception as e:
-            QMessageBox.critical(self, "Import Error", f"Error importing parameters: {str(e)}")
-    
-    def set_main_system_params(self, params):
-        """Set the main system parameters from a tuple or list"""
-        if not params or len(params) < 14:  # Check if we have at least the minimum required params
-            return False
-            
-        try:
-            # Set mu parameter
-            self.mu_box.setValue(float(params[0]))
-            
-            # Set lambda parameters
-            for i, value in enumerate(params[1:6]):
-                if i < len(self.landa_boxes):
-                    self.landa_boxes[i].setValue(float(value))
-            
-            # Set nu parameters
-            for i, value in enumerate(params[6:11]):
-                if i < len(self.nu_boxes):
-                    self.nu_boxes[i].setValue(float(value))
-            
-            # Set remaining parameters
-            self.a_low_box.setValue(float(params[11]))
-            self.a_up_box.setValue(float(params[12]))
-            self.f_1_box.setValue(float(params[13]))
-            
-            if len(params) >= 15:
-                self.f_2_box.setValue(float(params[14]))
-            if len(params) >= 16:
-                self.omega_dc_box.setValue(float(params[15]))
-            if len(params) >= 17:
-                self.zeta_dc_box.setValue(float(params[16]))
-                
-            return True
-        except (ValueError, IndexError) as e:
-            QMessageBox.warning(self, "Parameter Error", f"Error setting main system parameters: {str(e)}")
-            return False
-    
-    def set_dva_params(self, params):
-        """Set the DVA parameters from a dictionary"""
-        if not params:
-            return False
-            
-        try:
-            # Set beta parameters
-            if "beta" in params and isinstance(params["beta"], list):
-                for i, value in enumerate(params["beta"]):
-                    if i < len(self.beta_boxes):
-                        self.beta_boxes[i].setValue(float(value))
-            
-            # Set lambda parameters
-            if "lambda" in params and isinstance(params["lambda"], list):
-                for i, value in enumerate(params["lambda"]):
-                    if i < len(self.lambda_boxes):
-                        self.lambda_boxes[i].setValue(float(value))
-            
-            # Set mu parameters
-            if "mu" in params and isinstance(params["mu"], list):
-                for i, value in enumerate(params["mu"]):
-                    if i < len(self.mu_dva_boxes):
-                        self.mu_dva_boxes[i].setValue(float(value))
-            
-            # Set nu parameters
-            if "nu" in params and isinstance(params["nu"], list):
-                for i, value in enumerate(params["nu"]):
-                    if i < len(self.nu_dva_boxes):
-                        self.nu_dva_boxes[i].setValue(float(value))
-                        
-            return True
-        except (ValueError, IndexError) as e:
-            QMessageBox.warning(self, "Parameter Error", f"Error setting DVA parameters: {str(e)}")
-            return False
-
-    def get_num_samples_list(self):
-        """Get the list of sample sizes for Sobol analysis"""
-        num_samples_text = self.num_samples_line.text().strip()
-        if not num_samples_text:
-            return [32, 64, 128]  # Default values
-        
-        try:
-            # Parse comma-separated values
-            samples = [int(n.strip()) for n in num_samples_text.split(',') if n.strip()]
-            if not samples:
-                return [32, 64, 128]  # Default if parsing yields empty list
-            return samples
-        except ValueError:
-            QMessageBox.warning(self, "Input Error", 
-                               "Invalid num_samples format. Using default values: 32, 64, 128")
-            return [32, 64, 128]
-    
-    def apply_optimized_dva_parameters(self):
-        """Apply the optimized DVA parameters from the selected optimization method"""
-        selected_optimizer = self.dva_optimizer_combo.currentText()
-        
-        # Get appropriate parameters based on selection
-        best_params = None
-        best_fitness = None
-        
-        if "Genetic Algorithm" in selected_optimizer:
-            if hasattr(self, 'current_ga_best_params') and self.current_ga_best_params is not None:
-                best_params = self.current_ga_best_params
-                best_fitness = self.current_ga_best_fitness
-            else:
-                QMessageBox.warning(self, "No Data Available", 
-                    "No optimized GA parameters available. Please run GA optimization first.")
-                return
-        
-        elif "Particle Swarm" in selected_optimizer:
-            if hasattr(self, 'current_pso_best_params') and self.current_pso_best_params is not None:
-                best_params = self.current_pso_best_params
-                best_fitness = getattr(self, 'current_pso_best_fitness', None)
-            else:
-                QMessageBox.warning(self, "No Data Available", 
-                    "No optimized PSO parameters available. Please run PSO optimization first.")
-                return
-                
-        elif "Differential Evolution" in selected_optimizer:
-            if hasattr(self, 'current_de_best_params') and self.current_de_best_params is not None:
-                best_params = self.current_de_best_params
-                best_fitness = getattr(self, 'current_de_best_fitness', None)
-            else:
-                QMessageBox.warning(self, "No Data Available", 
-                    "No optimized DE parameters available. Please run DE optimization first.")
-                return
-                
-        elif "Simulated Annealing" in selected_optimizer:
-            if hasattr(self, 'current_sa_best_params') and self.current_sa_best_params is not None:
-                best_params = self.current_sa_best_params
-                best_fitness = getattr(self, 'current_sa_best_fitness', None)
-            else:
-                QMessageBox.warning(self, "No Data Available", 
-                    "No optimized SA parameters available. Please run SA optimization first.")
-                return
-                
-        elif "CMA-ES" in selected_optimizer:
-            if hasattr(self, 'current_cmaes_best_params') and self.current_cmaes_best_params is not None:
-                best_params = self.current_cmaes_best_params
-                best_fitness = getattr(self, 'current_cmaes_best_fitness', None)
-            else:
-                QMessageBox.warning(self, "No Data Available", 
-                    "No optimized CMA-ES parameters available. Please run CMA-ES optimization first.")
-                return
-                
-        elif "Reinforcement Learning" in selected_optimizer:
-            if hasattr(self, 'current_rl_best_params') and self.current_rl_best_params is not None:
-                best_params = self.current_rl_best_params
-                best_fitness = getattr(self, 'current_rl_best_fitness', None)
-            else:
-                QMessageBox.warning(self, "No Data Available", 
-                    "No optimized RL parameters available. Please run RL optimization first.")
-                return
-        
-        # If we have valid parameters, apply them to the DVA parameter inputs
-        if best_params:
-            # Apply parameters to the appropriate input fields
-            self.set_dva_params(best_params)
-            
-            # Show confirmation with fitness value if available
-            fitness_msg = f" (Fitness: {best_fitness:.6f})" if best_fitness is not None else ""
-            QMessageBox.information(self, "Parameters Applied", 
-                f"The optimized parameters from {selected_optimizer} have been applied to the DVA inputs.{fitness_msg}")
-        else:
-            QMessageBox.warning(self, "No Data Available", 
-                "No optimized parameters available from the selected method.")
-            
     def get_main_system_params(self):
         """Get the main system parameters from the UI"""
         return (
